CRUK footer

CRUK footer

Usage

crukFooter()

Value

a div containing a styled footer with links and logos

Examples

crukFooter()
#> <div class="central-column">
#>   <div class="info-footer">
#>     <div>
#>       <a href="https://www.cancerresearchuk.org/">
#>         <img src="shinyCRUK/images/cruk-logo.svg" class="cruk-logo-footer"/>
#>       </a>
#>     </div>
#>     <div class="bslib-grid bslib-mb-spacing html-fill-item" data-require-bs-caller="layout_column_wrap()" data-require-bs-version="5" style="grid-template-columns:repeat(2, minmax(0, 1fr));grid-auto-rows:1fr;--bslib-grid-height:auto;--bslib-grid-height-mobile:auto;">
#>       <div class="bslib-grid-item bslib-gap-spacing html-fill-container">
#> 
#> 
#> <p>You are welcome to reuse this Cancer Research UK-produced content for
#> your own work. Please remember to acknowledge the original data sources
#> for each chart and ensure you retain the CRUK logo in the image. Please
#> note that due to differences in data collection, comparison of
#> performance should not be made between nations.</p></div>
#>       <div class="bslib-grid-item bslib-gap-spacing html-fill-container">
#> 
#> 
#> <p>If you have any questions regarding the data or the interpretation of
#> this data, or if you are having any issues with the website please
#> contact:<br />
#> <span style="word-wrap: anywhere;"><a href="mailto:cancerintelligence@cancer.org.uk" class="email">cancerintelligence@cancer.org.uk</a></span> (please allow
#> up to 10 working days for a response)</p>
#> <p><a href="https://www.cancerresearchuk.org/health-professional/cancer-statistics-for-the-uk" class="stats-page">See our complete statistics pages on
#> cancerresearch.org</a></p></div>
#>     </div>
#>   </div>
#>   <hr class="footer-hr"/>
#>   <h3 class="footer-strapline">
#>     Together we are
#>     <br/>
#>      beating cancer
#>   </h3>
#>   <div class="terms-and-conditions">
#>     <div>
#>       <ul class="terms-and-conditions-list">
#>         <li>
#>           <a class="terms-and-conditions-link" href="https://www.cancerresearchuk.org/terms-and-conditions">Terms and conditions</a>
#>         </li>
#>         <li>
#>           <a class="terms-and-conditions-link" href="https://www.cancerresearchuk.org/about-us/our-organisation/privacy-statement">Privacy</a>
#>         </li>
#>         <li>
#>           <a class="terms-and-conditions-link" href="https://www.cancerresearchuk.org/about-us/our-organisation/responsible-organisation">Modern slavery statement</a>
#>         </li>
#>         <li>
#>           <a class="terms-and-conditions-link" href="https://crukcancerintelligence.shinyapps.io/cookiespolicy">Cookies</a>
#>         </li>
#>       </ul>
#>     </div>
#>   </div>
#>   <a href="https://www.fundraisingregulator.org.uk/">
#>     <img class="regulator-logo" src="shinyCRUK/images/logo-fundraising-regulator-reg.svg" height="97px"/>
#>   </a>
#>   <address>Cancer Research UK is a registered charity in England and Wales (1089464), Scotland (SC041666), the Isle of Man (1103) and Jersey (247). A company limited by guarantee. Registered company in England and Wales (4325234) and the Isle of Man (5713F). Registered address: 2 Redman Place, London, E20 1JQ.</address>
#> </div>

if (interactive()) {
  crukFooter()
}