Health Professional Last Review

Health Professional Last Review

Usage

hpReview(last_review_date)

Arguments

last_review_date

A date to displayed as the last date the page or content was reviewed

Value

A div containing a styled health professional box and the date of last review

Examples

hpReview("08 July 2025")
#> <div style="margin-bottom: 8px;&#10;             display: flex;">
#>   <span style="background: #e2f4fd;&#10;                  color: #0076b4;&#10;                  font-family: Poppins;&#10;                  display: inline-flex;&#10;                  align-items: center;&#10;                  padding: 8px 12px 8px 12px;">Health professionals</span>
#>   <div style="flex: 1;&#10;                   align-self: stretch;"></div>
#>   <span style="color: #666666;&#10;                  font-family: Poppins;&#10;                  padding: 8px 12px 8px 12px;&#10;                  text-align: right;">Last reviewed: 08 July 2025</span>
#> </div>