CRUK footer

CRUK footer

Usage

crukFooter()

Value

a div containing a styled footer with links and logos

Examples

crukFooter()
#> <div class="info-footer">
#>   <div>
#>     <a href="https://www.cancerresearchuk.org/">
#>       <img src="data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMjY4IiBoZWlnaHQ9IjEwOCIgdmlld0JveD0iMCAwIDI2OCAxMDgiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+CjxwYXRoIGQ9Ik0zNy41NTc1IDc1LjA1NjlDNDEuNTExOSA3NS4wNTY5IDQ0LjcxNzcgNzEuNzc2NyA0NC43MTc3IDY3LjczMDJDNDQuNzE3NyA2My42ODM4IDQxLjUxMTkgNjAuNDAzNiAzNy41NTc1IDYwLjQwMzZDMzMuNjAzMSA2MC40MDM2IDMwLjM5NzMgNjMuNjgzOCAzMC4zOTczIDY3LjczMDJDMzAuMzk3MyA3MS43NzY3IDMzLjYwMzEgNzUuMDU2OSAzNy41NTc1IDc1LjA1NjlaIiBmaWxsPSIjMDAwMDdFIi8+CjxwYXRoIGQ9Ik0xNi41NjgyIDI1LjAwMjJDMTguMDkwMiAyNS4wMDIyIDE5LjMyNCAyMy43NDIzIDE5LjMyNCAyMi4xODgxQzE5LjMyNCAyMC42MzM5IDE4LjA5MDIgMTkuMzc0IDE2LjU2ODIgMTkuMzc0QzE1LjA0NjIgMTkuMzc0IDEzLjgxMjQgMjAuNjMzOSAxMy44MTI0IDIyLjE4ODFDMTMuODEyNCAyMy43NDIzIDE1LjA0NjIgMjUuMDAyMiAxNi41NjgyIDI1LjAwMjJaIiBmaWxsPSIjRkYwMDg3Ii8+CjxwYXRoIGQ9Ik0yOS44NDc5IDU5LjgzNzJDMzEuMzY5OSA1OS44MzcyIDMyLjYwMzcgNTguNTc3MyAzMi42MDM3IDU3LjAyMzFDMzIuNjAzNyA1NS40Njg5IDMxLjM2OTkgNTQuMjA5IDI5Ljg0NzkgNTQuMjA5QzI4LjMyNTkgNTQuMjA5IDI3LjA5MiA1NS40Njg5IDI3LjA5MiA1Ny4wMjMxQzI3LjA5MiA1OC41NzczIDI4LjMyNTkgNTkuODM3MiAyOS44NDc5IDU5LjgzNzJaIiBmaWxsPSIjRkYwMDg3Ii8+CjxwYXRoIGQ9Ik0yNi40Njc2IDY3LjgyMThDMjcuOTg5NiA2Ny44MjE4IDI5LjIyMzUgNjYuNTYxOSAyOS4yMjM1IDY1LjAwNzdDMjkuMjIzNSA2My40NTM1IDI3Ljk4OTYgNjIuMTkzNiAyNi40Njc2IDYyLjE5MzZDMjQuOTQ1NiA2Mi4xOTM2IDIzLjcxMTggNjMuNDUzNSAyMy43MTE4IDY1LjAwNzdDMjMuNzExOCA2Ni41NjE5IDI0Ljk0NTYgNjcuODIxOCAyNi40Njc2IDY3LjgyMThaIiBmaWxsPSIjRkYwMDg3Ii8+CjxwYXRoIGQ9Ik0yNy42NDk5IDk4LjU1MjFDMjkuMTcxOSA5OC41NTIxIDMwLjQwNTcgOTcuMjkyMSAzMC40MDU3IDk1LjczNzlDMzAuNDA1NyA5NC4xODM4IDI5LjE3MTkgOTIuOTIzOCAyNy42NDk5IDkyLjkyMzhDMjYuMTI3OSA5Mi45MjM4IDI0Ljg5NCA5NC4xODM4IDI0Ljg5NCA5NS43Mzc5QzI0Ljg5NCA5Ny4yOTIxIDI2LjEyNzkgOTguNTUyMSAyNy42NDk5IDk4LjU1MjFaIiBmaWxsPSIjMDA5Q0VFIi8+CjxwYXRoIGQ9Ik03MS40ODUxIDEwNy40OTRDNzMuMDA3MSAxMDcuNDk0IDc0LjI0MDkgMTA2LjIzNCA3NC4yNDA5IDEwNC42OEM3NC4yNDA5IDEwMy4xMjYgNzMuMDA3MSAxMDEuODY2IDcxLjQ4NTEgMTAxLjg2NkM2OS45NjMxIDEwMS44NjYgNjguNzI5MiAxMDMuMTI2IDY4LjcyOTIgMTA0LjY4QzY4LjcyOTIgMTA2LjIzNCA2OS45NjMxIDEwNy40OTQgNzEuNDg1MSAxMDcuNDk0WiIgZmlsbD0iI0ZGMDA4NyIvPgo8cGF0aCBkPSJNODMuMzk5MyA5My4wMDcxQzg0LjkyMTMgOTMuMDA3MSA4Ni4xNTUxIDkxLjc0NzIgODYuMTU1MSA5MC4xOTNDODYuMTU1MSA4OC42Mzg4IDg0LjkyMTMgODcuMzc4OSA4My4zOTkzIDg3LjM3ODlDODEuODc3MyA4Ny4zNzg5IDgwLjY0MzQgODguNjM4OCA4MC42NDM0IDkwLjE5M0M4MC42NDM0IDkxLjc0NzIgODEuODc3MyA5My4wMDcxIDgzLjM5OTMgOTMuMDA3MVoiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTc4LjY4NjcgMTA2LjM0NUM4MC4yMDg3IDEwNi4zNDUgODEuNDQyNiAxMDUuMDg1IDgxLjQ0MjYgMTAzLjUzMUM4MS40NDI2IDEwMS45NzcgODAuMjA4NyAxMDAuNzE3IDc4LjY4NjcgMTAwLjcxN0M3Ny4xNjQ3IDEwMC43MTcgNzUuOTMwOSAxMDEuOTc3IDc1LjkzMDkgMTAzLjUzMUM3NS45MzA5IDEwNS4wODUgNzcuMTY0NyAxMDYuMzQ1IDc4LjY4NjcgMTA2LjM0NVoiIGZpbGw9IiMwMDlDRUUiLz4KPHBhdGggZD0iTTkwLjM2NzkgODguNDk0N0M5MS44ODk5IDg4LjQ5NDcgOTMuMTIzNyA4Ny4yMzQ4IDkzLjEyMzcgODUuNjgwNkM5My4xMjM3IDg0LjEyNjQgOTEuODg5OSA4Mi44NjY1IDkwLjM2NzkgODIuODY2NUM4OC44NDU5IDgyLjg2NjUgODcuNjEyMSA4NC4xMjY0IDg3LjYxMjEgODUuNjgwNkM4Ny42MTIxIDg3LjIzNDggODguODQ1OSA4OC40OTQ3IDkwLjM2NzkgODguNDk0N1oiIGZpbGw9IiMwMDlDRUUiLz4KPHBhdGggZD0iTTM2LjM2NjkgMjkuMjk4NEMzNy44ODg5IDI5LjI5ODQgMzkuMTIyNyAyOC4wMzg1IDM5LjEyMjcgMjYuNDg0M0MzOS4xMjI3IDI0LjkzMDEgMzcuODg4OSAyMy42NzAyIDM2LjM2NjkgMjMuNjcwMkMzNC44NDQ5IDIzLjY3MDIgMzMuNjExMSAyNC45MzAxIDMzLjYxMTEgMjYuNDg0M0MzMy42MTExIDI4LjAzODUgMzQuODQ0OSAyOS4yOTg0IDM2LjM2NjkgMjkuMjk4NFoiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTQ3LjQ1NjkgODEuOTgzOUM1MC4wOTE3IDgxLjk4MzkgNTIuMjI3NiA3OS43OTU4IDUyLjIyNzYgNzcuMDk2N0M1Mi4yMjc2IDc0LjM5NzYgNTAuMDkxNyA3Mi4yMDk1IDQ3LjQ1NjkgNzIuMjA5NUM0NC44MjIyIDcyLjIwOTUgNDIuNjg2MyA3NC4zOTc2IDQyLjY4NjMgNzcuMDk2N0M0Mi42ODYzIDc5Ljc5NTggNDQuODIyMiA4MS45ODM5IDQ3LjQ1NjkgODEuOTgzOVoiIGZpbGw9IiMwMDAwN0UiLz4KPHBhdGggZD0iTTIxLjA3MjUgNjIuMzEwMUMyMy43MDczIDYyLjMxMDEgMjUuODQzMiA2MC4xMjIgMjUuODQzMiA1Ny40MjI5QzI1Ljg0MzIgNTQuNzIzNyAyMy43MDczIDUyLjUzNTYgMjEuMDcyNSA1Mi41MzU2QzE4LjQzNzggNTIuNTM1NiAxNi4zMDE5IDU0LjcyMzcgMTYuMzAxOSA1Ny40MjI5QzE2LjMwMTkgNjAuMTIyIDE4LjQzNzggNjIuMzEwMSAyMS4wNzI1IDYyLjMxMDFaIiBmaWxsPSIjMDA5Q0VFIi8+CjxwYXRoIGQ9Ik04Mi42NjY3IDg1LjUyMjNDODUuMzAxNCA4NS41MjIzIDg3LjQzNzMgODMuMzM0MiA4Ny40MzczIDgwLjYzNUM4Ny40MzczIDc3LjkzNTkgODUuMzAxNCA3NS43NDc4IDgyLjY2NjcgNzUuNzQ3OEM4MC4wMzE5IDc1Ljc0NzggNzcuODk2IDc3LjkzNTkgNzcuODk2IDgwLjYzNUM3Ny44OTYgODMuMzM0MiA4MC4wMzE5IDg1LjUyMjMgODIuNjY2NyA4NS41MjIzWiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNNzUuMjE1IDk5LjY4NDRDNzcuODQ5OCA5OS42ODQ0IDc5Ljk4NTcgOTcuNDk2MyA3OS45ODU3IDk0Ljc5NzFDNzkuOTg1NyA5Mi4wOTggNzcuODQ5OCA4OS45MDk5IDc1LjIxNSA4OS45MDk5QzcyLjU4MDIgODkuOTA5OSA3MC40NDQzIDkyLjA5OCA3MC40NDQzIDk0Ljc5NzFDNzAuNDQ0MyA5Ny40OTYzIDcyLjU4MDIgOTkuNjg0NCA3NS4yMTUgOTkuNjg0NFoiIGZpbGw9IiMwMDAwN0UiLz4KPHBhdGggZD0iTTg2LjgyMTIgMTAzLjc3MkM4OS40NTYgMTAzLjc3MiA5MS41OTE5IDEwMS41ODQgOTEuNTkxOSA5OC44ODUzQzkxLjU5MTkgOTYuMTg2MSA4OS40NTYgOTMuOTk4IDg2LjgyMTIgOTMuOTk4Qzg0LjE4NjQgOTMuOTk4IDgyLjA1MDUgOTYuMTg2MSA4Mi4wNTA1IDk4Ljg4NTNDODIuMDUwNSAxMDEuNTg0IDg0LjE4NjQgMTAzLjc3MiA4Ni44MjEyIDEwMy43NzJaIiBmaWxsPSIjMDAwMDdFIi8+CjxwYXRoIGQ9Ik05OS4xMzQ5IDg2LjM3OTlDMTAxLjc3IDg2LjM3OTkgMTAzLjkwNiA4NC4xOTE4IDEwMy45MDYgODEuNDkyN0MxMDMuOTA2IDc4Ljc5MzYgMTAxLjc3IDc2LjYwNTUgOTkuMTM0OSA3Ni42MDU1Qzk2LjUwMDIgNzYuNjA1NSA5NC4zNjQzIDc4Ljc5MzYgOTQuMzY0MyA4MS40OTI3Qzk0LjM2NDMgODQuMTkxOCA5Ni41MDAyIDg2LjM3OTkgOTkuMTM0OSA4Ni4zNzk5WiIgZmlsbD0iI0ZGMDA4NyIvPgo8cGF0aCBkPSJNNjIuNjY4MSAxMDYuODk1QzY1LjMwMjkgMTA2Ljg5NSA2Ny40Mzg4IDEwNC43MDYgNjcuNDM4OCAxMDIuMDA3QzY3LjQzODggOTkuMzA4MiA2NS4zMDI5IDk3LjEyMDEgNjIuNjY4MSA5Ny4xMjAxQzYwLjAzMzQgOTcuMTIwMSA1Ny44OTc1IDk5LjMwODIgNTcuODk3NSAxMDIuMDA3QzU3Ljg5NzUgMTA0LjcwNiA2MC4wMzM0IDEwNi44OTUgNjIuNjY4MSAxMDYuODk1WiIgZmlsbD0iIzAwOUNFRSIvPgo8cGF0aCBkPSJNNTEuNTE5OSAxMDMuNTIyQzU0LjE1NDcgMTAzLjUyMiA1Ni4yOTA2IDEwMS4zMzQgNTYuMjkwNiA5OC42MzUzQzU2LjI5MDYgOTUuOTM2MSA1NC4xNTQ3IDkzLjc0OCA1MS41MTk5IDkzLjc0OEM0OC44ODUyIDkzLjc0OCA0Ni43NDkzIDk1LjkzNjEgNDYuNzQ5MyA5OC42MzUzQzQ2Ljc0OTMgMTAxLjMzNCA0OC44ODUyIDEwMy41MjIgNTEuNTE5OSAxMDMuNTIyWiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNMjYuMDUxNCA4MC44NTE2QzI4LjY4NjIgODAuODUxNiAzMC44MjIxIDc4LjY2MzUgMzAuODIyMSA3NS45NjQ0QzMwLjgyMjEgNzMuMjY1MiAyOC42ODYyIDcxLjA3NzEgMjYuMDUxNCA3MS4wNzcxQzIzLjQxNjcgNzEuMDc3MSAyMS4yODA4IDczLjI2NTIgMjEuMjgwOCA3NS45NjQ0QzIxLjI4MDggNzguNjYzNSAyMy40MTY3IDgwLjg1MTYgMjYuMDUxNCA4MC44NTE2WiIgZmlsbD0iIzAwOUNFRSIvPgo8cGF0aCBkPSJNMi43NTU4MyAzOC44MzEzQzQuMjc3ODMgMzguODMxMyA1LjUxMTY2IDM3LjU3MTQgNS41MTE2NiAzNi4wMTcyQzUuNTExNjYgMzQuNDYzIDQuMjc3ODMgMzMuMjAzMSAyLjc1NTgzIDMzLjIwMzFDMS4yMzM4MyAzMy4yMDMxIDAgMzQuNDYzIDAgMzYuMDE3MkMwIDM3LjU3MTQgMS4yMzM4MyAzOC44MzEzIDIuNzU1ODMgMzguODMxM1oiIGZpbGw9IiMwMDlDRUUiLz4KPHBhdGggZD0iTTguMjU5MTIgNjEuNjEwNkM5Ljc4MTEzIDYxLjYxMDYgMTEuMDE1IDYwLjM1MDcgMTEuMDE1IDU4Ljc5NjVDMTEuMDE1IDU3LjI0MjMgOS43ODExMyA1NS45ODI0IDguMjU5MTIgNTUuOTgyNEM2LjczNzEyIDU1Ljk4MjQgNS41MDMzIDU3LjI0MjMgNS41MDMzIDU4Ljc5NjVDNS41MDMzIDYwLjM1MDcgNi43MzcxMiA2MS42MTA2IDguMjU5MTIgNjEuNjEwNloiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTM1LjAzNDkgODIuNzc0N0MzNi41NTY5IDgyLjc3NDcgMzcuNzkwNyA4MS41MTQ4IDM3Ljc5MDcgNzkuOTYwNkMzNy43OTA3IDc4LjQwNjQgMzYuNTU2OSA3Ny4xNDY1IDM1LjAzNDkgNzcuMTQ2NUMzMy41MTI5IDc3LjE0NjUgMzIuMjc5MSA3OC40MDY0IDMyLjI3OTEgNzkuOTYwNkMzMi4yNzkxIDgxLjUxNDggMzMuNTEyOSA4Mi43NzQ3IDM1LjAzNDkgODIuNzc0N1oiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTE4LjQwODMgODYuMjc5OEMxOS45MzAzIDg2LjI3OTggMjEuMTY0MSA4NS4wMTk5IDIxLjE2NDEgODMuNDY1N0MyMS4xNjQxIDgxLjkxMTUgMTkuOTMwMyA4MC42NTE2IDE4LjQwODMgODAuNjUxNkMxNi44ODYzIDgwLjY1MTYgMTUuNjUyNSA4MS45MTE1IDE1LjY1MjUgODMuNDY1N0MxNS42NTI1IDg1LjAxOTkgMTYuODg2MyA4Ni4yNzk4IDE4LjQwODMgODYuMjc5OFoiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTUxLjg2OTUgOTEuNzE2NkM1My4zOTE1IDkxLjcxNjYgNTQuNjI1MyA5MC40NTY3IDU0LjYyNTMgODguOTAyNUM1NC42MjUzIDg3LjM0ODMgNTMuMzkxNSA4Ni4wODg0IDUxLjg2OTUgODYuMDg4NEM1MC4zNDc1IDg2LjA4ODQgNDkuMTEzNiA4Ny4zNDgzIDQ5LjExMzYgODguOTAyNUM0OS4xMTM2IDkwLjQ1NjcgNTAuMzQ3NSA5MS43MTY2IDUxLjg2OTUgOTEuNzE2NloiIGZpbGw9IiMwMDlDRUUiLz4KPHBhdGggZD0iTTU0Ljk1IDg0LjU4OTlDNTYuNDcyMSA4NC41ODk5IDU3LjcwNTkgODMuMzMgNTcuNzA1OSA4MS43NzU4QzU3LjcwNTkgODAuMjIxNiA1Ni40NzIxIDc4Ljk2MTcgNTQuOTUgNzguOTYxN0M1My40MjggNzguOTYxNyA1Mi4xOTQyIDgwLjIyMTYgNTIuMTk0MiA4MS43NzU4QzUyLjE5NDIgODMuMzMgNTMuNDI4IDg0LjU4OTkgNTQuOTUgODQuNTg5OVoiIGZpbGw9IiMwMDAwN0UiLz4KPHBhdGggZD0iTTc0LjIzMjUgODcuMDIxQzc1Ljc1NDUgODcuMDIxIDc2Ljk4ODMgODUuNzYxMSA3Ni45ODgzIDg0LjIwNjlDNzYuOTg4MyA4Mi42NTI3IDc1Ljc1NDUgODEuMzkyOCA3NC4yMzI1IDgxLjM5MjhDNzIuNzEwNSA4MS4zOTI4IDcxLjQ3NjcgODIuNjUyNyA3MS40NzY3IDg0LjIwNjlDNzEuNDc2NyA4NS43NjExIDcyLjcxMDUgODcuMDIxIDc0LjIzMjUgODcuMDIxWiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNMTA1LjI1NCA5MS4wMDA4QzEwNi43NzYgOTEuMDAwOCAxMDguMDEgODkuNzQwOSAxMDguMDEgODguMTg2N0MxMDguMDEgODYuNjMyNSAxMDYuNzc2IDg1LjM3MjYgMTA1LjI1NCA4NS4zNzI2QzEwMy43MzIgODUuMzcyNiAxMDIuNDk5IDg2LjYzMjUgMTAyLjQ5OSA4OC4xODY3QzEwMi40OTkgODkuNzQwOSAxMDMuNzMyIDkxLjAwMDggMTA1LjI1NCA5MS4wMDA4WiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNOTUuMjk2OCAxOS40NzM5Qzk3LjkzMTYgMTkuNDczOSAxMDAuMDY3IDE3LjI4NTggMTAwLjA2NyAxNC41ODY3QzEwMC4wNjcgMTEuODg3NSA5Ny45MzE2IDkuNjk5NDYgOTUuMjk2OCA5LjY5OTQ2QzkyLjY2MiA5LjY5OTQ2IDkwLjUyNjEgMTEuODg3NSA5MC41MjYxIDE0LjU4NjdDOTAuNTI2MSAxNy4yODU4IDkyLjY2MiAxOS40NzM5IDk1LjI5NjggMTkuNDczOVoiIGZpbGw9IiMwMDAwN0UiLz4KPHBhdGggZD0iTTEwMy40NzMgMjcuMDc1NUMxMDYuMTA3IDI3LjA3NTUgMTA4LjI0MyAyNC44ODc0IDEwOC4yNDMgMjIuMTg4M0MxMDguMjQzIDE5LjQ4OTEgMTA2LjEwNyAxNy4zMDEgMTAzLjQ3MyAxNy4zMDFDMTAwLjgzOCAxNy4zMDEgOTguNzAyIDE5LjQ4OTEgOTguNzAyIDIyLjE4ODNDOTguNzAyIDI0Ljg4NzQgMTAwLjgzOCAyNy4wNzU1IDEwMy40NzMgMjcuMDc1NVoiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTg4LjA5NSAyNy44NjYzQzkwLjcyOTggMjcuODY2MyA5Mi44NjU3IDI1LjY3ODIgOTIuODY1NyAyMi45NzlDOTIuODY1NyAyMC4yNzk5IDkwLjcyOTggMTguMDkxOCA4OC4wOTUgMTguMDkxOEM4NS40NjAyIDE4LjA5MTggODMuMzI0MyAyMC4yNzk5IDgzLjMyNDMgMjIuOTc5QzgzLjMyNDMgMjUuNjc4MiA4NS40NjAyIDI3Ljg2NjMgODguMDk1IDI3Ljg2NjNaIiBmaWxsPSIjMDA5Q0VFIi8+CjxwYXRoIGQ9Ik0yNS4wNTIzIDM0LjY2ODVDMjcuNjg3IDM0LjY2ODUgMjkuODIyOSAzMi40ODA0IDI5LjgyMjkgMjkuNzgxM0MyOS44MjI5IDI3LjA4MjEgMjcuNjg3IDI0Ljg5NCAyNS4wNTIzIDI0Ljg5NEMyMi40MTc1IDI0Ljg5NCAyMC4yODE2IDI3LjA4MjEgMjAuMjgxNiAyOS43ODEzQzIwLjI4MTYgMzIuNDgwNCAyMi40MTc1IDM0LjY2ODUgMjUuMDUyMyAzNC42Njg1WiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNODQuMTA2OSAzMy43NjFDODUuNjI4OSAzMy43NjEgODYuODYyNyAzMi41MDExIDg2Ljg2MjcgMzAuOTQ2OUM4Ni44NjI3IDI5LjM5MjcgODUuNjI4OSAyOC4xMzI4IDg0LjEwNjkgMjguMTMyOEM4Mi41ODQ5IDI4LjEzMjggODEuMzUxMSAyOS4zOTI3IDgxLjM1MTEgMzAuOTQ2OUM4MS4zNTExIDMyLjUwMTEgODIuNTg0OSAzMy43NjEgODQuMTA2OSAzMy43NjFaIiBmaWxsPSIjRkYwMDg3Ii8+CjxwYXRoIGQ9Ik03Ny4zMjE0IDI5Ljk0NzhDNzkuOTU2MiAyOS45NDc4IDgyLjA5MjEgMjcuNzU5NyA4Mi4wOTIxIDI1LjA2MDZDODIuMDkyMSAyMi4zNjE0IDc5Ljk1NjIgMjAuMTczMyA3Ny4zMjE0IDIwLjE3MzNDNzQuNjg2NyAyMC4xNzMzIDcyLjU1MDggMjIuMzYxNCA3Mi41NTA4IDI1LjA2MDZDNzIuNTUwOCAyNy43NTk3IDc0LjY4NjcgMjkuOTQ3OCA3Ny4zMjE0IDI5Ljk0NzhaIiBmaWxsPSIjMDAwMDdFIi8+CjxwYXRoIGQ9Ik05MS43NjY2IDQwLjE4ODVDOTQuNDAxNCA0MC4xODg1IDk2LjUzNzMgMzguMDAwNCA5Ni41MzczIDM1LjMwMTNDOTYuNTM3MyAzMi42MDIxIDk0LjQwMTQgMzAuNDE0MSA5MS43NjY2IDMwLjQxNDFDODkuMTMxOSAzMC40MTQxIDg2Ljk5NiAzMi42MDIxIDg2Ljk5NiAzNS4zMDEzQzg2Ljk5NiAzOC4wMDA0IDg5LjEzMTkgNDAuMTg4NSA5MS43NjY2IDQwLjE4ODVaIiBmaWxsPSIjRkYwMDg3Ii8+CjxwYXRoIGQ9Ik0zOC4yNTY5IDU4Ljc5NjdDNDAuODkxNiA1OC43OTY3IDQzLjAyNzUgNTYuNjA4NiA0My4wMjc1IDUzLjkwOTRDNDMuMDI3NSA1MS4yMTAzIDQwLjg5MTYgNDkuMDIyMiAzOC4yNTY5IDQ5LjAyMjJDMzUuNjIyMSA0OS4wMjIyIDMzLjQ4NjIgNTEuMjEwMyAzMy40ODYyIDUzLjkwOTRDMzMuNDg2MiA1Ni42MDg2IDM1LjYyMjEgNTguNzk2NyAzOC4yNTY5IDU4Ljc5NjdaIiBmaWxsPSIjMDAwMDdFIi8+CjxwYXRoIGQ9Ik05Ni4xNzA5IDMwLjA4OTRDOTcuNjkyOSAzMC4wODk0IDk4LjkyNjcgMjguODI5NSA5OC45MjY3IDI3LjI3NTNDOTguOTI2NyAyNS43MjExIDk3LjY5MjkgMjQuNDYxMiA5Ni4xNzA5IDI0LjQ2MTJDOTQuNjQ4OSAyNC40NjEyIDkzLjQxNSAyNS43MjExIDkzLjQxNSAyNy4yNzUzQzkzLjQxNSAyOC44Mjk1IDk0LjY0ODkgMzAuMDg5NCA5Ni4xNzA5IDMwLjA4OTRaIiBmaWxsPSIjMDAwMDdFIi8+CjxwYXRoIGQ9Ik05Ni4yODc1IDk3LjgxMTFDOTguOTIyMyA5Ny44MTExIDEwMS4wNTggOTUuNjIzIDEwMS4wNTggOTIuOTIzOEMxMDEuMDU4IDkwLjIyNDcgOTguOTIyMyA4OC4wMzY2IDk2LjI4NzUgODguMDM2NkM5My42NTI3IDg4LjAzNjYgOTEuNTE2OCA5MC4yMjQ3IDkxLjUxNjggOTIuOTIzOEM5MS41MTY4IDk1LjYyMyA5My42NTI3IDk3LjgxMTEgOTYuMjg3NSA5Ny44MTExWiIgZmlsbD0iI0ZGMDA4NyIvPgo8cGF0aCBkPSJNMzkuOTA1NCA0Ny42NDgyQzQxLjQyNzQgNDcuNjQ4MiA0Mi42NjEyIDQ2LjM4ODMgNDIuNjYxMiA0NC44MzQxQzQyLjY2MTIgNDMuMjc5OSA0MS40Mjc0IDQyLjAyIDM5LjkwNTQgNDIuMDJDMzguMzgzNCA0Mi4wMiAzNy4xNDk1IDQzLjI3OTkgMzcuMTQ5NSA0NC44MzQxQzM3LjE0OTUgNDYuMzg4MyAzOC4zODM0IDQ3LjY0ODIgMzkuOTA1NCA0Ny42NDgyWiIgZmlsbD0iI0ZGMDA4NyIvPgo8cGF0aCBkPSJNNTEuMzk1IDMzLjYxOTRDNTIuOTE3IDMzLjYxOTQgNTQuMTUwOCAzMi4zNTk1IDU0LjE1MDggMzAuODA1M0M1NC4xNTA4IDI5LjI1MTEgNTIuOTE3IDI3Ljk5MTIgNTEuMzk1IDI3Ljk5MTJDNDkuODczIDI3Ljk5MTIgNDguNjM5MiAyOS4yNTExIDQ4LjYzOTIgMzAuODA1M0M0OC42MzkyIDMyLjM1OTUgNDkuODczIDMzLjYxOTQgNTEuMzk1IDMzLjYxOTRaIiBmaWxsPSIjMDAwMDdFIi8+CjxwYXRoIGQ9Ik01Mi43NzcxIDIwLjcwNjFDNTQuMjk5MSAyMC43MDYxIDU1LjUzMjkgMTkuNDQ2MiA1NS41MzI5IDE3Ljg5MkM1NS41MzI5IDE2LjMzNzggNTQuMjk5MSAxNS4wNzc5IDUyLjc3NzEgMTUuMDc3OUM1MS4yNTUxIDE1LjA3NzkgNTAuMDIxMiAxNi4zMzc4IDUwLjAyMTIgMTcuODkyQzUwLjAyMTIgMTkuNDQ2MiA1MS4yNTUxIDIwLjcwNjEgNTIuNzc3MSAyMC43MDYxWiIgZmlsbD0iI0ZGMDA4NyIvPgo8cGF0aCBkPSJNNS41MTE2OSA3Ny44NzkyQzcuMDMzNjkgNzcuODc5MiA4LjI2NzUyIDc2LjYxOTMgOC4yNjc1MiA3NS4wNjUxQzguMjY3NTIgNzMuNTEwOSA3LjAzMzY5IDcyLjI1MSA1LjUxMTY5IDcyLjI1MUMzLjk4OTY5IDcyLjI1MSAyLjc1NTg2IDczLjUxMDkgMi43NTU4NiA3NS4wNjUxQzIuNzU1ODYgNzYuNjE5MyAzLjk4OTY5IDc3Ljg3OTIgNS41MTE2OSA3Ny44NzkyWiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNMTguMzY2NyA3MC41NDQyQzE5Ljg4ODcgNzAuNTQ0MiAyMS4xMjI1IDY5LjI4NDMgMjEuMTIyNSA2Ny43MzAxQzIxLjEyMjUgNjYuMTc1OSAxOS44ODg3IDY0LjkxNiAxOC4zNjY3IDY0LjkxNkMxNi44NDQ3IDY0LjkxNiAxNS42MTA4IDY2LjE3NTkgMTUuNjEwOCA2Ny43MzAxQzE1LjYxMDggNjkuMjg0MyAxNi44NDQ3IDcwLjU0NDIgMTguMzY2NyA3MC41NDQyWiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNMjkuMjE1MiA4OC40OTQ3QzMwLjczNzIgODguNDk0NyAzMS45NzEgODcuMjM0OCAzMS45NzEgODUuNjgwNkMzMS45NzEgODQuMTI2NCAzMC43MzcyIDgyLjg2NjUgMjkuMjE1MiA4Mi44NjY1QzI3LjY5MzIgODIuODY2NSAyNi40NTk0IDg0LjEyNjQgMjYuNDU5NCA4NS42ODA2QzI2LjQ1OTQgODcuMjM0OCAyNy42OTMyIDg4LjQ5NDcgMjkuMjE1MiA4OC40OTQ3WiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNMTguMzY2NyA0MS4yOTU3QzE5Ljg4ODcgNDEuMjk1NyAyMS4xMjI1IDQwLjAzNTggMjEuMTIyNSAzOC40ODE2QzIxLjEyMjUgMzYuOTI3NCAxOS44ODg3IDM1LjY2NzUgMTguMzY2NyAzNS42Njc1QzE2Ljg0NDcgMzUuNjY3NSAxNS42MTA4IDM2LjkyNzQgMTUuNjEwOCAzOC40ODE2QzE1LjYxMDggNDAuMDM1OCAxNi44NDQ3IDQxLjI5NTcgMTguMzY2NyA0MS4yOTU3WiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNMzMuNjYxMSAzNi42NjY4QzM1LjE4MzEgMzYuNjY2OCAzNi40MTY5IDM1LjQwNjkgMzYuNDE2OSAzMy44NTI3QzM2LjQxNjkgMzIuMjk4NSAzNS4xODMxIDMxLjAzODYgMzMuNjYxMSAzMS4wMzg2QzMyLjEzOTEgMzEuMDM4NiAzMC45MDUzIDMyLjI5ODUgMzAuOTA1MyAzMy44NTI3QzMwLjkwNTMgMzUuNDA2OSAzMi4xMzkxIDM2LjY2NjggMzMuNjYxMSAzNi42NjY4WiIgZmlsbD0iIzAwOUNFRSIvPgo8cGF0aCBkPSJNMTUuNjE5MSA1MC40OTU5QzE3LjE0MTEgNTAuNDk1OSAxOC4zNzQ5IDQ5LjIzNiAxOC4zNzQ5IDQ3LjY4MThDMTguMzc0OSA0Ni4xMjc2IDE3LjE0MTEgNDQuODY3NyAxNS42MTkxIDQ0Ljg2NzdDMTQuMDk3MSA0NC44Njc3IDEyLjg2MzMgNDYuMTI3NiAxMi44NjMzIDQ3LjY4MThDMTIuODYzMyA0OS4yMzYgMTQuMDk3MSA1MC40OTU5IDE1LjYxOTEgNTAuNDk1OVoiIGZpbGw9IiMwMDlDRUUiLz4KPHBhdGggZD0iTTQyLjk0NDMgNDAuOTA0NkM0NS41NzkgNDAuOTA0NiA0Ny43MTQ5IDM4LjcxNjUgNDcuNzE0OSAzNi4wMTc0QzQ3LjcxNDkgMzMuMzE4MiA0NS41NzkgMzEuMTMwMSA0Mi45NDQzIDMxLjEzMDFDNDAuMzA5NSAzMS4xMzAxIDM4LjE3MzYgMzMuMzE4MiAzOC4xNzM2IDM2LjAxNzRDMzguMTczNiAzOC43MTY1IDQwLjMwOTUgNDAuOTA0NiA0Mi45NDQzIDQwLjkwNDZaIiBmaWxsPSIjMDAwMDdFIi8+CjxwYXRoIGQ9Ik0zMS42NjMgMjMuNjM2OEMzNC4yOTc4IDIzLjYzNjggMzYuNDMzNyAyMS40NDg3IDM2LjQzMzcgMTguNzQ5NUMzNi40MzM3IDE2LjA1MDQgMzQuMjk3OCAxMy44NjIzIDMxLjY2MyAxMy44NjIzQzI5LjAyODIgMTMuODYyMyAyNi44OTIzIDE2LjA1MDQgMjYuODkyMyAxOC43NDk1QzI2Ljg5MjMgMjEuNDQ4NyAyOS4wMjgyIDIzLjYzNjggMzEuNjYzIDIzLjYzNjhaIiBmaWxsPSIjMDA5Q0VFIi8+CjxwYXRoIGQ9Ik01OS4yODc4IDI5LjM0ODRDNjEuOTIyNSAyOS4zNDg0IDY0LjA1ODQgMjcuMTYwMyA2NC4wNTg0IDI0LjQ2MTJDNjQuMDU4NCAyMS43NjIxIDYxLjkyMjUgMTkuNTc0IDU5LjI4NzggMTkuNTc0QzU2LjY1MyAxOS41NzQgNTQuNTE3MSAyMS43NjIxIDU0LjUxNzEgMjQuNDYxMkM1NC41MTcxIDI3LjE2MDMgNTYuNjUzIDI5LjM0ODQgNTkuMjg3OCAyOS4zNDg0WiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNNjkuMzAzNyA5Ljc3NDQ1QzcxLjkzODUgOS43NzQ0NSA3NC4wNzQ0IDcuNTg2MzcgNzQuMDc0NCA0Ljg4NzIzQzc0LjA3NDQgMi4xODgwOCA3MS45Mzg1IDAgNjkuMzAzNyAwQzY2LjY2OSAwIDY0LjUzMzEgMi4xODgwOCA2NC41MzMxIDQuODg3MjNDNjQuNTMzMSA3LjU4NjM3IDY2LjY2OSA5Ljc3NDQ1IDY5LjMwMzcgOS43NzQ0NVoiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTcwLjc2OTEgMjAuNjQ4QzczLjQwMzggMjAuNjQ4IDc1LjUzOTcgMTguNDU5OSA3NS41Mzk3IDE1Ljc2MDhDNzUuNTM5NyAxMy4wNjE2IDczLjQwMzggMTAuODczNSA3MC43NjkxIDEwLjg3MzVDNjguMTM0MyAxMC44NzM1IDY1Ljk5ODQgMTMuMDYxNiA2NS45OTg0IDE1Ljc2MDhDNjUuOTk4NCAxOC40NTk5IDY4LjEzNDMgMjAuNjQ4IDcwLjc2OTEgMjAuNjQ4WiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNNTYuOTIzMiAxMC45NDg1QzU5LjU1OCAxMC45NDg1IDYxLjY5MzkgOC43NjA0NCA2MS42OTM5IDYuMDYxM0M2MS42OTM5IDMuMzYyMTYgNTkuNTU4IDEuMTc0MDcgNTYuOTIzMiAxLjE3NDA3QzU0LjI4ODUgMS4xNzQwNyA1Mi4xNTI2IDMuMzYyMTYgNTIuMTUyNiA2LjA2MTNDNTIuMTUyNiA4Ljc2MDQ0IDU0LjI4ODUgMTAuOTQ4NSA1Ni45MjMyIDEwLjk0ODVaIiBmaWxsPSIjMDA5Q0VFIi8+CjxwYXRoIGQ9Ik02OC4zMjEzIDI3LjYzMzFDNjkuODQzMyAyNy42MzMxIDcxLjA3NzEgMjYuMzczMiA3MS4wNzcxIDI0LjgxOUM3MS4wNzcxIDIzLjI2NDggNjkuODQzMyAyMi4wMDQ5IDY4LjMyMTMgMjIuMDA0OUM2Ni43OTkzIDIyLjAwNDkgNjUuNTY1NCAyMy4yNjQ4IDY1LjU2NTQgMjQuODE5QzY1LjU2NTQgMjYuMzczMiA2Ni43OTkzIDI3LjYzMzEgNjguMzIxMyAyNy42MzMxWiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNNjAuOTYxMyAxNy42NzU2QzYyLjQ4MzMgMTcuNjc1NiA2My43MTcxIDE2LjQxNTcgNjMuNzE3MSAxNC44NjE1QzYzLjcxNzEgMTMuMzA3MyA2Mi40ODMzIDEyLjA0NzQgNjAuOTYxMyAxMi4wNDc0QzU5LjQzOTMgMTIuMDQ3NCA1OC4yMDU0IDEzLjMwNzMgNTguMjA1NCAxNC44NjE1QzU4LjIwNTQgMTYuNDE1NyA1OS40MzkzIDE3LjY3NTYgNjAuOTYxMyAxNy42NzU2WiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNOTEuMzY3IDc4LjQwMzlDOTQuMDAxOCA3OC40MDM5IDk2LjEzNzYgNzYuMjE1OCA5Ni4xMzc2IDczLjUxNjZDOTYuMTM3NiA3MC44MTc1IDk0LjAwMTggNjguNjI5NCA5MS4zNjcgNjguNjI5NEM4OC43MzIyIDY4LjYyOTQgODYuNTk2MyA3MC44MTc1IDg2LjU5NjMgNzMuNTE2NkM4Ni41OTYzIDc2LjIxNTggODguNzMyMiA3OC40MDM5IDkxLjM2NyA3OC40MDM5WiIgZmlsbD0iIzAwMDA3RSIvPgo8cGF0aCBkPSJNNDUuMTY3MyAyOS4yODE4QzQ3LjgwMiAyOS4yODE4IDQ5LjkzNzkgMjcuMDkzNyA0OS45Mzc5IDI0LjM5NDVDNDkuOTM3OSAyMS42OTU0IDQ3LjgwMiAxOS41MDczIDQ1LjE2NzMgMTkuNTA3M0M0Mi41MzI1IDE5LjUwNzMgNDAuMzk2NiAyMS42OTU0IDQwLjM5NjYgMjQuMzk0NUM0MC4zOTY2IDI3LjA5MzcgNDIuNTMyNSAyOS4yODE4IDQ1LjE2NzMgMjkuMjgxOFoiIGZpbGw9IiMwMDAwN0UiLz4KPHBhdGggZD0iTTgzLjA5OTUgMTYuNTUxN0M4Ni43ODI2IDE2LjU1MTcgODkuNzY4NCAxMy40OTg4IDg5Ljc2ODQgOS43MzI4N0M4OS43Njg0IDUuOTY2OTQgODYuNzgyNiAyLjkxNDA2IDgzLjA5OTUgMi45MTQwNkM3OS40MTYzIDIuOTE0MDYgNzYuNDMwNSA1Ljk2Njk0IDc2LjQzMDUgOS43MzI4N0M3Ni40MzA1IDEzLjQ5ODggNzkuNDE2MyAxNi41NTE3IDgzLjA5OTUgMTYuNTUxN1oiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTQ0LjAxIDE3LjgxNzFDNDcuNjkzMiAxNy44MTcxIDUwLjY3ODkgMTQuNzY0MiA1MC42Nzg5IDEwLjk5ODJDNTAuNjc4OSA3LjIzMjMzIDQ3LjY5MzIgNC4xNzk0NCA0NC4wMSA0LjE3OTQ0QzQwLjMyNjkgNC4xNzk0NCAzNy4zNDExIDcuMjMyMzMgMzcuMzQxMSAxMC45OTgyQzM3LjM0MTEgMTQuNzY0MiA0MC4zMjY5IDE3LjgxNzEgNDQuMDEgMTcuODE3MVoiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTI4LjU3NDEgNTEuMDcwMkMzMi4yNTcyIDUxLjA3MDIgMzUuMjQzIDQ4LjAxNzMgMzUuMjQzIDQ0LjI1MTRDMzUuMjQzIDQwLjQ4NTUgMzIuMjU3MiAzNy40MzI2IDI4LjU3NDEgMzcuNDMyNkMyNC44OTA5IDM3LjQzMjYgMjEuOTA1MiA0MC40ODU1IDIxLjkwNTIgNDQuMjUxNEMyMS45MDUyIDQ4LjAxNzMgMjQuODkwOSA1MS4wNzAyIDI4LjU3NDEgNTEuMDcwMloiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTQwLjUzODIgOTcuMTAzNEM0NC4yMjE0IDk3LjEwMzQgNDcuMjA3MSA5NC4wNTA1IDQ3LjIwNzEgOTAuMjg0NkM0Ny4yMDcxIDg2LjUxODcgNDQuMjIxNCA4My40NjU4IDQwLjUzODIgODMuNDY1OEMzNi44NTUgODMuNDY1OCAzMy44NjkzIDg2LjUxODcgMzMuODY5MyA5MC4yODQ2QzMzLjg2OTMgOTQuMDUwNSAzNi44NTUgOTcuMTAzNCA0MC41MzgyIDk3LjEwMzRaIiBmaWxsPSIjRkYwMDg3Ii8+CjxwYXRoIGQ9Ik02NC4wMDAxIDk0LjIyMjhDNjcuNjgzMyA5NC4yMjI4IDcwLjY2OTEgOTEuMTY5OSA3MC42NjkxIDg3LjQwNEM3MC42NjkxIDgzLjYzODEgNjcuNjgzMyA4MC41ODUyIDY0LjAwMDEgODAuNTg1MkM2MC4zMTcgODAuNTg1MiA1Ny4zMzEyIDgzLjYzODEgNTcuMzMxMiA4Ny40MDRDNTcuMzMxMiA5MS4xNjk5IDYwLjMxNyA5NC4yMjI4IDY0LjAwMDEgOTQuMjIyOFoiIGZpbGw9IiNGRjAwODciLz4KPHBhdGggZD0iTTEzOC45OTkgMzEuNzQ2MUMxMzcuNzkxIDMzLjI3ODEgMTM1LjgzNSAzNC4yMjcyIDEzMy44NzggMzQuMjI3MkMxMzEuOTIyIDM0LjIyNzIgMTMwLjMyMyAzMy41MTk1IDEyOS4wOTEgMzIuMjI5QzEyNy44NDIgMzAuOTMwMiAxMjcuMTU5IDI5LjEwNjkgMTI3LjE1OSAyNy4xMDg3QzEyNy4xNTkgMjUuMTEwNSAxMjcuODQyIDIzLjI4NzEgMTI5LjA5MSAyMS45ODgzQzEzMC4zMjMgMjAuNzA2MiAxMzIuMDIyIDE5Ljk5MDEgMTMzLjg3OCAxOS45OTAxQzEzNS43MzUgMTkuOTkwMSAxMzcuNzY2IDIwLjkzOTMgMTM4Ljk5OSAyMi40NzEyTDE0MS4zODggMjAuMDgxN0MxMzkuNTU2IDE3Ljk3NTMgMTM2Ljc5MiAxNi43NjgxIDEzMy43OTUgMTYuNzY4MUMxMzAuOTU2IDE2Ljc2ODEgMTI4LjM1OCAxNy44MTcxIDEyNi40NjggMTkuNzE1NEMxMjQuNTc4IDIxLjYyMiAxMjMuNTM4IDI0LjI0NDYgMTIzLjUzOCAyNy4xMDg3QzEyMy41MzggMjkuOTcyNyAxMjQuNTc4IDMyLjU5NTQgMTI2LjQ2OCAzNC41MDJDMTI4LjM1OCAzNi40MDAyIDEzMC45NzMgMzcuNDQ5MyAxMzMuODIgMzcuNDQ5M0MxMzYuODI2IDM3LjQ0OTMgMTM5LjU4MSAzNi4yNDIgMTQxLjM4OCAzNC4xMzU2TDEzOC45OTkgMzEuNzQ2MVoiIGZpbGw9ImJsYWNrIi8+CjxwYXRoIGQ9Ik0xNzUuOTY1IDM3LjExNjJIMTc5LjY2MlYxNy4xMTc3SDE3Ni4zODFWMzEuMDg4M0wxNjguMTMxIDE3LjExNzdIMTY0LjI0MlYzNy4xMTYySDE2Ny40ODlWMjIuNzIwOUwxNzUuOTY1IDM3LjExNjJaIiBmaWxsPSJibGFjayIvPgo8cGF0aCBkPSJNMTk4LjgxOSAzMS43NDYxQzE5Ny42MTIgMzMuMjc4MSAxOTUuNjU2IDM0LjIyNzIgMTkzLjY5OSAzNC4yMjcyQzE5MS43NDIgMzQuMjI3MiAxOTAuMTQ0IDMzLjUxOTUgMTg4LjkxMiAzMi4yMjlDMTg3LjY2MyAzMC45MzAyIDE4Ni45OCAyOS4xMDY5IDE4Ni45OCAyNy4xMDg3QzE4Ni45OCAyNS4xMTA1IDE4Ny42NjMgMjMuMjg3MSAxODguOTEyIDIxLjk4ODNDMTkwLjE0NCAyMC43MDYyIDE5MS44NDIgMTkuOTkwMSAxOTMuNjk5IDE5Ljk5MDFDMTk1LjU1NiAxOS45OTAxIDE5Ny41ODcgMjAuOTM5MyAxOTguODE5IDIyLjQ3MTJMMjAxLjIwOSAyMC4wODE3QzE5OS4zNzcgMTcuOTc1MyAxOTYuNjEzIDE2Ljc2ODEgMTkzLjYxNiAxNi43NjgxQzE5MC43NzcgMTYuNzY4MSAxODguMTc5IDE3LjgxNzEgMTg2LjI4OSAxOS43MTU0QzE4NC4zOTkgMjEuNjIyIDE4My4zNTggMjQuMjQ0NiAxODMuMzU4IDI3LjEwODdDMTgzLjM1OCAyOS45NzI3IDE4NC4zOTkgMzIuNTk1NCAxODYuMjg5IDM0LjUwMkMxODguMTc5IDM2LjQwMDIgMTkwLjc5MyAzNy40NDkzIDE5My42NDEgMzcuNDQ5M0MxOTYuNjQ2IDM3LjQ0OTMgMTk5LjQwMiAzNi4yNDIgMjAxLjIwOSAzNC4xMzU2TDE5OC44MTkgMzEuNzQ2MVoiIGZpbGw9ImJsYWNrIi8+CjxwYXRoIGQ9Ik0yMDguMDYxIDIwLjExNUgyMTcuMTExVjE3LjExNzdIMjA0LjYzOVYzNy4xMTYySDIxNy4xMTFWMzQuMTE4OUgyMDguMDYxVjI4LjQ0MDdIMjE1LjY4N1YyNS40NTE4SDIwOC4wNjFWMjAuMTE1WiIgZmlsbD0iYmxhY2siLz4KPHBhdGggZD0iTTEzMC45MzEgNTYuODRMMTM1LjU5MyA2NC4wOTE4SDEzOS41NzNMMTM0LjQ5NCA1Ni40OTAzTDEzNC41NDQgNTYuNDczN0MxMzcuMTg0IDU1LjYyNDQgMTM4Ljg5IDUzLjI2ODMgMTM4Ljg5IDUwLjQ3OTFDMTM4Ljg5IDQ2Ljc3NDIgMTM1Ljk2OCA0NC4wOTMzIDEzMS45MyA0NC4wOTMzSDEyNC42NDVWNjQuMDkxOEgxMjguMDY3VjU2Ljg0SDEzMC45MzFaTTEyOC4wNjcgNDcuMDA3M0gxMzEuMjgxQzEzMy44MTIgNDcuMDA3MyAxMzUuMjY5IDQ4LjI3MjggMTM1LjI2OSA1MC40ODc1QzEzNS4yNjkgNTIuNzAyMSAxMzMuNzg3IDUzLjkzNDMgMTMxLjMwNiA1My45MzQzSDEyOC4wNjdWNDcuMDA3M1oiIGZpbGw9ImJsYWNrIi8+CjxwYXRoIGQ9Ik0yMjcuNjY4IDI5LjgxNDZMMjMyLjMzMSAzNy4wNjY0SDIzNi4zMUwyMzEuMjMyIDI5LjQ2NDlMMjMxLjI4MiAyOS40NDgzQzIzMy45MjEgMjguNTk5MSAyMzUuNjI4IDI2LjI0MjkgMjM1LjYyOCAyMy40NTM3QzIzNS42MjggMTkuNzQ4OCAyMzIuNzA1IDE3LjA2NzkgMjI4LjY2NyAxNy4wNjc5SDIyMS4zODJWMzcuMDY2NEgyMjQuODA0VjI5LjgxNDZIMjI3LjY2OFpNMjI0LjgwNCAxOS45ODE5SDIyOC4wMThDMjMwLjU0OSAxOS45ODE5IDIzMi4wMDYgMjEuMjQ3NCAyMzIuMDA2IDIzLjQ2MjFDMjMyLjAwNiAyNS42NzY3IDIzMC41MjQgMjYuOTA4OSAyMjguMDQzIDI2LjkwODlIMjI0LjgwNFYxOS45ODE5WiIgZmlsbD0iYmxhY2siLz4KPHBhdGggZD0iTTIxOS4xNTkgNTYuODRMMjIzLjgyMiA2NC4wOTE4SDIyNy44MDFMMjIyLjcyMyA1Ni40OTAzTDIyMi43NzIgNTYuNDczN0MyMjUuNDEyIDU1LjYyNDQgMjI3LjExOSA1My4yNjgzIDIyNy4xMTkgNTAuNDc5MUMyMjcuMTE5IDQ2Ljc3NDIgMjI0LjE5NiA0NC4wOTMzIDIyMC4xNTggNDQuMDkzM0gyMTIuODczVjY0LjA5MThIMjE2LjI5NVY1Ni44NEgyMTkuMTU5Wk0yMTYuMjk1IDQ3LjAwNzNIMjE5LjUwOUMyMjIuMDQgNDcuMDA3MyAyMjMuNDk3IDQ4LjI3MjggMjIzLjQ5NyA1MC40ODc1QzIyMy40OTcgNTIuNzAyMSAyMjIuMDE1IDUzLjkzNDMgMjE5LjUzNCA1My45MzQzSDIxNi4yOTVWNDcuMDA3M1oiIGZpbGw9ImJsYWNrIi8+CjxwYXRoIGQ9Ik0xNDYuOTE3IDQ3LjA4MjJIMTU1Ljk3NVY0NC4wODVIMTQzLjQ5NVY2NC4wODM1SDE1NS45NzVWNjEuMDg2MkgxNDYuOTE3VjU1LjQxNjNIMTU0LjU1MVY1Mi40MTkxSDE0Ni45MTdWNDcuMDgyMloiIGZpbGw9ImJsYWNrIi8+CjxwYXRoIGQ9Ik0xNjYuMjY2IDY0LjQyNDlDMTcwLjQwMyA2NC40MjQ5IDE3My4wNzYgNjIuMDkzNiAxNzMuMDc2IDU4LjQ4ODZDMTczLjA3NiA1NS43NjYxIDE3MS40MzYgNTMuNzc2MiAxNjguNDcyIDUyLjg5MzdMMTY2LjE0MSA1Mi4xODZDMTY0LjEwOSA1MS41OTQ5IDE2My4yNDMgNTAuNzIwNiAxNjMuMjQzIDQ5LjI4MDNDMTYzLjI0MyA0Ny43MzE3IDE2NC41MTcgNDYuNjU3NyAxNjYuMzQ5IDQ2LjY1NzdDMTY3Ljg2NCA0Ni42NTc3IDE2OS4wOTYgNDcuMjMyMiAxNzAuMzI4IDQ4LjUxNDNMMTcyLjU1MSA0Ni4yOTEzQzE3MS4wNTMgNDQuNjI2MiAxNjguOTIxIDQzLjc0MzcgMTY2LjQwNyA0My43NDM3QzE2Mi40MDIgNDMuNzQzNyAxNTkuNzA1IDQ2LjE0OTggMTU5LjcwNSA0OS43Mjk5QzE1OS43MDUgNTEuMDEyMSAxNjAuMTA0IDUyLjE0NDQgMTYwLjg5NSA1My4wNzY4QzE2MS42OTUgNTQuMDI2IDE2Mi44NTIgNTQuNzMzNyAxNjQuMzQyIDU1LjE4MzNMMTY3LjEzMSA1Ni4wMDc1QzE2OC43MjIgNTYuNDgyMSAxNjkuNjA0IDU3LjQ4MTIgMTY5LjYwNCA1OC43OTY2QzE2OS42MDQgNjAuNDcwMSAxNjguMzMgNjEuNTAyNSAxNjYuMjc0IDYxLjUwMjVDMTY0LjQ3NSA2MS41MDI1IDE2Mi44OTQgNjAuNjY5OSAxNjEuNDM3IDU4Ljk2MzJMMTU5LjE4OSA2MS4yMTExQzE2MC45MDQgNjMuMjQyNiAxNjMuNDg1IDY0LjQxNjUgMTY2LjI3NCA2NC40MTY1TDE2Ni4yNjYgNjQuNDI0OVoiIGZpbGw9ImJsYWNrIi8+CjxwYXRoIGQ9Ik0xODAuMjg2IDQ3LjA4MjJIMTg5LjMzNlY0NC4wODVIMTc2Ljg1NlY2NC4wODM1SDE4OS4zMzZWNjEuMDg2MkgxODAuMjg2VjU1LjQxNjNIMTg3LjkxM1Y1Mi40MTkxSDE4MC4yODZWNDcuMDgyMloiIGZpbGw9ImJsYWNrIi8+CjxwYXRoIGQ9Ik0xOTcuMDYzIDU5LjU2MjZIMjA0LjU2NEwyMDYuMDcxIDY0LjA4MzVIMjA5LjgzNEwyMDIuOTgyIDQ0LjA4NUgxOTguODg2TDE5MiA2NC4wODM1SDE5NS41MzFMMTk3LjA2MyA1OS41NjI2Wk0yMDAuODQyIDQ4LjMxNDVMMjAzLjY0OCA1Ni43NDAxSDE5OC4wMDNMMjAwLjg0MiA0OC4zMTQ1WiIgZmlsbD0iYmxhY2siLz4KPHBhdGggZD0iTTE0OC4yOSAzMi41OTUzSDE1NS43OTJMMTU3LjI5OSAzNy4xMTYySDE2MS4wNjJMMTU0LjIxIDE3LjExNzdIMTUwLjExNEwxNDMuMjI4IDM3LjExNjJIMTQ2Ljc1OEwxNDguMjkgMzIuNTk1M1pNMTUyLjA3IDIxLjM0NzJMMTU0Ljg3NiAyOS43NzI4SDE0OS4yMzFMMTUyLjA3IDIxLjM0NzJaIiBmaWxsPSJibGFjayIvPgo8cGF0aCBkPSJNMjQ1LjUyNyA0OS40NTUxTDI0Ny45MTYgNDcuMDY1NkMyNDYuMDg1IDQ0Ljk1OTIgMjQzLjMyMSA0My43NTIgMjQwLjMyMyA0My43NTJDMjM3LjQ4NCA0My43NTIgMjM0Ljg4NyA0NC44MDEgMjMyLjk5NyA0Ni42OTkzQzIzMS4xMDcgNDguNjA1OSAyMzAuMDY2IDUxLjIyODUgMjMwLjA2NiA1NC4wOTI2QzIzMC4wNjYgNTYuOTU2NiAyMzEuMTA3IDU5LjU3OTIgMjMyLjk5NyA2MS40ODU4QzIzNC44ODcgNjMuMzg0MSAyMzcuNTAxIDY0LjQzMzIgMjQwLjM0OCA2NC40MzMyQzI0My4zNTQgNjQuNDMzMiAyNDYuMTEgNjMuMjI1OSAyNDcuOTE2IDYxLjExOTVMMjQ1LjUyNyA1OC43M0MyNDQuMzIgNjAuMjYyIDI0Mi4zNjMgNjEuMjExMSAyNDAuNDA3IDYxLjIxMTFDMjM4LjQ1IDYxLjIxMTEgMjM2Ljg1MSA2MC41MDM0IDIzNS42MTkgNTkuMjEyOUMyMzQuMzcgNTcuOTE0MSAyMzMuNjg4IDU2LjA5MDcgMjMzLjY4OCA1NC4wOTI2QzIzMy42ODggNTIuMDk0NCAyMzQuMzcgNTAuMjcxIDIzNS42MTkgNDguOTcyMkMyMzYuODUxIDQ3LjY5IDIzOC41NSA0Ni45NzQgMjQwLjQwNyA0Ni45NzRDMjQyLjI2MyA0Ni45NzQgMjQ0LjI5NSA0Ny45MjMyIDI0NS41MjcgNDkuNDU1MVoiIGZpbGw9ImJsYWNrIi8+CjxwYXRoIGQ9Ik0yNjMuODI3IDY0LjA4MzVIMjY3LjI3NFY0NC4wODVIMjYzLjgyN1Y1Mi4zNzc0VjUyLjQxOTFIMjYzLjc4NUgyNTUuMTUySDI1NS4xMVY1Mi4zNzc0VjQ0LjA4NUgyNTEuNjg4VjY0LjA4MzVIMjU1LjExVjU1LjQ0OTZWNTUuNDE2M0gyNTUuMTUySDI2My43ODVIMjYzLjgyN1Y1NS40NDk2VjY0LjA4MzVaIiBmaWxsPSJibGFjayIvPgo8cGF0aCBkPSJNMTMyLjA4OCA5MS42OTE3QzEzNi44NTEgOTEuNjkxNyAxMzkuNTQ4IDg5LjA1MjQgMTM5LjU0OCA4NC42MjMxVjcxLjM1MThIMTM2LjEyNlY4NC4zOUMxMzYuMTEgODUuNjgwNSAxMzUuNzQzIDg2Ljc2MjggMTM1LjAxMSA4Ny41MDM4QzEzNC4yNzggODguMjQ0OCAxMzMuMzcgODguNjM2MSAxMzIuMDg4IDg4LjYzNjFDMTMwLjgwNiA4OC42MzYxIDEyOS44ODIgODguMjQ0OCAxMjkuMTQ5IDg3LjUwMzhDMTI4LjQyNSA4Ni43NjI4IDEyOC4wMzQgODUuNjg4OCAxMjguMDM0IDg0LjM5VjcxLjM1MThIMTI0LjYxMlY4NC42MjMxQzEyNC42MTIgODkuMDUyNCAxMjcuMzM0IDkxLjY5MTcgMTMyLjA4OCA5MS42OTE3WiIgZmlsbD0iYmxhY2siLz4KPHBhdGggZD0iTTE0OS4zMTQgODIuMDkyTDE1Ni4xMjUgOTEuMzUwM0gxNjAuMjk2TDE1MS43MDQgNzkuNjk0MkwxNTkuOTk2IDcxLjM1MThIMTU1LjQzNEwxNDcgODAuMTAyMlY3MS4zNTE4SDE0My41NzhWOTEuMzUwM0gxNDdWODQuNDMxNkwxNDkuMzE0IDgyLjA5MloiIGZpbGw9ImJsYWNrIi8+Cjwvc3ZnPgo=" class="cruk-logo-footer"/>
#>     </a>
#>   </div>
#>   <div class="bslib-grid bslib-mb-spacing html-fill-item" data-require-bs-caller="layout_column_wrap()" data-require-bs-version="5" style="grid-template-columns:repeat(2, minmax(0, 1fr));grid-auto-rows:1fr;--bslib-grid-height:auto;--bslib-grid-height-mobile:auto;">
#>     <div class="bslib-grid-item bslib-gap-spacing html-fill-container">
#> 
#> 
#> <p>You are welcome to reuse this Cancer Research UK-produced content for
#> your own work. Please remember to acknowledge the original data sources
#> for each chart and ensure you retain the CRUK logo in the image. Please
#> note that due to differences in data collection, comparison of
#> performance should not be made between nations.</p></div>
#>     <div class="bslib-grid-item bslib-gap-spacing html-fill-container">
#> 
#> 
#> <p>If you have any questions regarding the data or the interpretation of
#> this data, or if you are having any issues with the website please
#> contact:<br />
#> <span style="word-wrap: anywhere;"><a href="mailto:cancerintelligence@cancer.org.uk" class="email">cancerintelligence@cancer.org.uk</a></span> (please allow
#> up to 10 working days for a response)</p>
#> <p><a href="https://www.cancerresearchuk.org/health-professional/cancer-statistics-for-the-uk" class="stats-page">See our complete statistics pages on
#> cancerresearch.org</a></p></div>
#>   </div>
#> </div>
#> <hr class="footer-hr"/>
#> <h3 class="footer-strapline">
#>   Together we are
#>   <br/>
#>    beating cancer
#> </h3>
#> <div class="terms-and-conditions">
#>   <div>
#>     <ul class="terms-and-conditions-list">
#>       <li>
#>         <a class="terms-and-conditions-link" href="https://www.cancerresearchuk.org/terms-and-conditions">Terms and conditions</a>
#>       </li>
#>       <li>
#>         <a class="terms-and-conditions-link" href="https://www.cancerresearchuk.org/about-us/our-organisation/privacy-statement">Privacy</a>
#>       </li>
#>       <li>
#>         <a class="terms-and-conditions-link" href="https://www.cancerresearchuk.org/about-us/our-organisation/responsible-organisation">Modern slavery statement</a>
#>       </li>
#>       <li>
#>         <a class="terms-and-conditions-link" href="https://crukcancerintelligence.shinyapps.io/cookiespolicy">Cookies</a>
#>       </li>
#>     </ul>
#>   </div>
#> </div>
#> <a href="https://www.fundraisingregulator.org.uk/">
#>   <img class="regulator-logo" src="data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTY1IiBoZWlnaHQ9IjUxIiB2aWV3Qm94PSIwIDAgMTY1IDUxIiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8ZyBjbGlwLXBhdGg9InVybCgjY2xpcDBfNDQ0Nl81MjczKSI+CjxwYXRoIGQ9Ik0yNS4yNzcyIDQuMTkxMjhDMzYuOTI2MiA0LjE5MTI4IDQ2LjM5OTcgMTMuNzQ4MyA0Ni4zOTk3IDI1LjVDNDYuMzk5NyAzNy4yNTE3IDM2LjkyNjIgNDYuODA4NyAyNS4yNzcyIDQ2LjgwODdDMTMuNjI4MSA0Ni44MDg3IDQuMTU0NjUgMzcuMjQ0NCA0LjE1NDY1IDI1LjVDNC4xNTQ2NSAxMy43NTU2IDEzLjYyODEgNC4xOTEyOCAyNS4yNzcyIDQuMTkxMjhaTTI1LjI3NzIgMEMxMS4zMTQzIDAgMCAxMS40MTQxIDAgMjUuNUMwIDM5LjU4NTkgMTEuMzE0MyA1MSAyNS4yNzcyIDUxQzM5LjI0IDUxIDUwLjU1NDMgMzkuNTg1OSA1MC41NTQzIDI1LjVDNTAuNTU0MyAxMS40MTQxIDM5LjIzMjcgMCAyNS4yNzcyIDBaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik0xMC4yNTkyIDE2Ljk0ODZIMjQuNjE1VjIxLjM0NTRIMTUuMjI4OFYyNC40MjFIMjMuNzEyN1YyOC41ODI5SDE1LjIyODhWMzUuMDQ5N0gxMC4yNTJWMTYuOTQ4NkgxMC4yNTkyWiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNMjUuNjExNyAxNi45NDg2SDM0LjA5NTZDMzYuODM4NyAxNi45NDg2IDM4LjczNzggMTcuNjc1MyAzOS45Mzg0IDE4Ljg4NjRDNDAuOTg2MSAxOS45NDM0IDQxLjUyNDUgMjEuMzE2MSA0MS41MjQ1IDIzLjA5OTdWMjMuMTUxMUM0MS41MjQ1IDI1Ljk0NzggNDAuMDYyIDI3Ljc4MjggMzcuODM1NiAyOC43MzdMNDIuMTEzOSAzNS4wNDk3SDM2LjM5NDlMMzIuNzc4NyAyOS41NjY1SDMwLjU3NFYzNS4wNDk3SDI1LjU5NzJWMTYuOTQ4NkgyNS42MTE3Wk0zMy44OTE5IDI1LjYzOTVDMzUuNTU4MSAyNS42Mzk1IDM2LjU1NSAyNC44MSAzNi41NTUgMjMuNDg4OFYyMy40Mzc0QzM2LjU1NSAyMS45OTE0IDM1LjUwNzIgMjEuMjY0NyAzMy44NjI4IDIxLjI2NDdIMzAuNTgxM1YyNS42Mzk1SDMzLjg5MTlaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik02MC4zMTg4IDE3LjczNEg2OC4xMDQzVjE4LjQ5NzRINjEuMTQxVjIzLjA3MDRINjcuMzk4NVYyMy44MzM4SDYxLjE0MVYyOC44ODM5SDYwLjMxODhWMTcuNzM0WiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNNjkuNzE5NyAyNC4yNTIyVjE3LjczNEg3MC41NDE5VjI0LjE3MTRDNzAuNTQxOSAyNi43OTkyIDcxLjk0NjIgMjguMzExMyA3NC4yODE4IDI4LjMxMTNDNzYuNjE3NSAyOC4zMTEzIDc3Ljk2MzUgMjYuOTM4NyA3Ny45NjM1IDI0LjI0NDhWMTcuNzI2N0g3OC43ODU3VjI0LjE0OTRDNzguNzg1NyAyNy4zNDk3IDc2LjkzNzYgMjkuMDUyNyA3NC4yNTI3IDI5LjA1MjdDNzEuNTY3OSAyOS4wNTI3IDY5LjcxOTcgMjcuMzY0NCA2OS43MTk3IDI0LjIzNzVWMjQuMjUyMloiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTgxLjEwNjQgMTcuNzM0SDgxLjg3NzdMODkuNTAzIDI3LjQzNzhWMTcuNzM0SDkwLjI5NjFWMjguODgzOUg4OS42Nzc2TDgxLjg5MjIgMTguOTc0NVYyOC44ODM5SDgxLjA5OTFWMTcuNzM0SDgxLjEwNjRaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik05Mi43OTkzIDE3LjczNEg5Ni40ODFDOTkuOTUxNyAxNy43MzQgMTAyLjM1MyAyMC4xNDE2IDEwMi4zNTMgMjMuMjc1OVYyMy4zMDUzQzEwMi4zNTMgMjYuNDQ2OSA5OS45NTE3IDI4Ljg4MzkgOTYuNDgxIDI4Ljg4MzlIOTIuNzk5M1YxNy43MzRaTTk2LjQ4MSAyOC4xMjA1Qzk5LjU1ODggMjguMTIwNSAxMDEuNTAyIDI1Ljk5OTEgMTAxLjUwMiAyMy4zNDJWMjMuMzEyNkMxMDEuNTAyIDIwLjY3MDEgOTkuNTU4OCAxOC41MDQ3IDk2LjQ4MSAxOC41MDQ3SDkzLjYyMTVWMjguMTI3OEg5Ni40ODFWMjguMTIwNVoiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTEwNC4zNjEgMTcuNzM0SDEwOS4wMDNDMTEwLjM2NCAxNy43MzQgMTExLjQ1NSAxOC4xNjcxIDExMi4xMTcgMTguODM1MUMxMTIuNjQxIDE5LjM2MzYgMTEyLjk1NCAyMC4xMTIzIDExMi45NTQgMjAuOTA1VjIwLjkzNDRDMTEyLjk1NCAyMi43OTg4IDExMS41NzkgMjMuODYzMSAxMDkuNzE2IDI0LjEyTDExMy4zNjEgMjguODgzOUgxMTIuMzIxTDEwOC43OTkgMjQuMjY2OEgxMDUuMTgzVjI4Ljg4MzlIMTA0LjM2MVYxNy43MzRaTTEwOC45MDggMjMuNTE4MUMxMTAuNzU3IDIzLjUxODEgMTEyLjEzMiAyMi41Nzg2IDExMi4xMzIgMjAuOTcxMVYyMC45NDE3QzExMi4xMzIgMTkuNDQ0MyAxMTAuOTYgMTguNTA0NyAxMDguOTU5IDE4LjUwNDdIMTA1LjE4M1YyMy41MjU1SDEwOC45MDhWMjMuNTE4MVoiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTExOS4xOTcgMTcuNjUzM0gxMTkuOTlMMTI1LjEzNCAyOC44ODM5SDEyNC4yMzJMMTIyLjgyOCAyNS43NjQySDExNi4zMDhMMTE0LjkwNCAyOC44ODM5SDExNC4wNTNMMTE5LjIwNCAxNy42NTMzSDExOS4xOTdaTTEyMi40ODYgMjUuMDE1NUwxMTkuNTgzIDE4LjU2MzVMMTE2LjY2NSAyNS4wMTU1SDEyMi40OTNIMTIyLjQ4NloiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTEyNi43OTMgMTcuNzM0SDEyNy42MTVWMjguODgzOUgxMjYuNzkzVjE3LjczNFoiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTEyOS41MTUgMjcuMjYxN0wxMzAuMDUzIDI2LjYzNzdDMTMxLjI2OCAyNy43NjgxIDEzMi4zODkgMjguMjk2NiAxMzQuMDE5IDI4LjI5NjZDMTM1LjY0OCAyOC4yOTY2IDEzNi44MTMgMjcuMzU3MSAxMzYuODEzIDI2LjA3OTlWMjYuMDUwNUMxMzYuODEzIDI0Ljg2ODcgMTM2LjE5NCAyNC4xODYxIDEzMy42ODQgMjMuNjc5NkMxMzEuMDI4IDIzLjEzNjQgMTI5Ljk1OCAyMi4yMjYzIDEyOS45NTggMjAuNTM4VjIwLjUwODZDMTI5Ljk1OCAxOC44NDk3IDEzMS40NTcgMTcuNTc5OSAxMzMuNTE3IDE3LjU3OTlDMTM1LjEyNSAxNy41Nzk5IDEzNi4xODcgMTguMDI3NiAxMzcuMzA3IDE4LjkzNzhMMTM2Ljc4MyAxOS41OTExQzEzNS43NDMgMTguNjgwOSAxMzQuNjk1IDE4LjMzNTkgMTMzLjQ4IDE4LjMzNTlDMTMxLjg1IDE4LjMzNTkgMTMwLjc4MSAxOS4yNzU1IDEzMC43ODEgMjAuNDM1MlYyMC40NjQ2QzEzMC43ODEgMjEuNjQ2NCAxMzEuMzc3IDIyLjM1ODQgMTM0LjAwNCAyMi45MDE2QzEzNi41NjUgMjMuNDMwMSAxMzcuNjM1IDI0LjM0NzYgMTM3LjYzNSAyNS45NzcxVjI2LjAwNjVDMTM3LjYzNSAyNy44MDQ4IDEzNi4xIDI5LjA1MjcgMTMzLjk2OCAyOS4wNTI3QzEzMi4yIDI5LjA1MjcgMTMwLjgzOSAyOC40NjU1IDEyOS41MTUgMjcuMjY5VjI3LjI2MTdaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik0xMzkuNjUgMTcuNzM0SDE0MC40NzJWMjguODgzOUgxMzkuNjVWMTcuNzM0WiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNMTQzLjA4NCAxNy43MzRIMTQzLjg1NkwxNTEuNDgxIDI3LjQzNzhWMTcuNzM0SDE1Mi4yNzRWMjguODgzOUgxNTEuNjU2TDE0My44NyAxOC45NzQ1VjI4Ljg4MzlIMTQzLjA3N1YxNy43MzRIMTQzLjA4NFoiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTE1NC4yODIgMjMuMzQyVjIzLjMxMjZDMTU0LjI4MiAyMC4zMDMxIDE1Ni40NDMgMTcuNTQzMiAxNTkuNzQ3IDE3LjU0MzJDMTYxLjU0NCAxNy41NDMyIDE2Mi42ODYgMTguMDg2NCAxNjMuNzkyIDE5LjAxMTJMMTYzLjI1NCAxOS42MzUxQzE2Mi4zODggMTguODcxOCAxNjEuMzc2IDE4LjI5OTIgMTU5LjcwMyAxOC4yOTkyQzE1Ny4wMTggMTguMjk5MiAxNTUuMTQxIDIwLjU5NjcgMTU1LjE0MSAyMy4yODMyVjIzLjMxMjZDMTU1LjE0MSAyNi4xOTczIDE1Ni45MjMgMjguMzQ4IDE1OS44NzggMjguMzQ4QzE2MS4yODIgMjguMzQ4IDE2Mi41OTIgMjcuNzYwOCAxNjMuMzk5IDI3LjA3MDhWMjMuODcwNUgxNTkuNzAzVjIzLjEwNzFIMTY0LjE4NVYyNy40MDg1QzE2My4xODggMjguMzA0IDE2MS42NDYgMjkuMDgyIDE1OS44NDEgMjkuMDgyQzE1Ni4zNDkgMjkuMDgyIDE1NC4yODIgMjYuNDgzNiAxNTQuMjgyIDIzLjM0OTNWMjMuMzQyWiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNNjAuMTY2IDMxLjY4NzhINjUuODM0MUM2Ny40MTMgMzEuNjg3OCA2OC42MjgxIDMyLjEzNTYgNjkuNDQzIDMyLjk1NzdDNzAuMTM0MyAzMy42NTUgNzAuNTA1MyAzNC42Mzg2IDcwLjUwNTMgMzUuODEzVjM1Ljg0OTdDNzAuNTA1MyAzNy44NjgzIDY5LjQyODUgMzkuMTM4MiA2Ny44NDk2IDM5LjcyNTRMNzAuODc2NCA0NC4xODgzSDY3LjY4OTVMNjUuMDMzNyA0MC4xODc4SDYyLjg5NDZWNDQuMTg4M0g2MC4xNjZWMzEuNjg3OFpNNjUuNjUyMiAzNy43NTgyQzY2Ljk3NjQgMzcuNzU4MiA2Ny43NDA0IDM3LjA0NjIgNjcuNzQwNCAzNS45ODkyVjM1Ljk1MjVDNjcuNzQwNCAzNC43NzA3IDY2LjkyNTUgMzQuMTY4OCA2NS42MDEzIDM0LjE2ODhINjIuODk0NlYzNy43NTgySDY1LjY1OTVINjUuNjUyMloiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTcyLjA2OTggMzEuNjg3OEg4MS40MTk2VjM0LjEzMjFINzQuNzgzOFYzNi42NzE4SDgwLjYyNjVWMzkuMTE2MUg3NC43ODM4VjQxLjc0NEg4MS41MTQyVjQ0LjE4ODNINzIuMDc3MVYzMS42ODc4SDcyLjA2OThaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik04Mi40ODkzIDM3Ljk3MTFWMzcuOTM0NEM4Mi40ODkzIDM0LjM4MTcgODUuMjMyMyAzMS40Njc2IDg4Ljk4NjggMzEuNDY3NkM5MS4yMjA2IDMxLjQ2NzYgOTIuNTY2NyAzMi4wNzY5IDkzLjg1NDUgMzMuMTg1Mkw5Mi4xMzc0IDM1LjI3NzJDOTEuMTg0MiAzNC40Njk4IDkwLjMzMjkgMzQuMDA3MyA4OC44OTk1IDM0LjAwNzNDODYuOTEzMSAzNC4wMDczIDg1LjM0MTUgMzUuNzc2MyA4NS4zNDE1IDM3Ljg5NzdWMzcuOTM0NEM4NS4zNDE1IDQwLjIxNzIgODYuODk4NiA0MS44OTgxIDg5LjA5NiA0MS44OTgxQzkwLjA4NTUgNDEuODk4MSA5MC45NzMyIDQxLjY0ODUgOTEuNjY0NCA0MS4xNDk0VjM5LjM2NTdIODguOTIxM1YzNi45OTQ4SDk0LjMwNTZWNDIuNDI2NkM5My4wMzIzIDQzLjUxMyA5MS4yNzg4IDQ0LjQwODUgODkuMDA4NiA0NC40MDg1Qzg1LjE1MjMgNDQuNDA4NSA4Mi40OTY1IDQxLjY3NzkgODIuNDk2NSAzNy45Nzg0TDgyLjQ4OTMgMzcuOTcxMVoiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTk1LjgzMzUgMzguODI5OVYzMS42ODc4SDk4LjU2MlYzOC43NTY1Qzk4LjU2MiA0MC43ODk3IDk5LjU3MzQgNDEuODQ2NyAxMDEuMjMyIDQxLjg0NjdDMTAyLjg5MSA0MS44NDY3IDEwMy45MDMgNDAuODI2NCAxMDMuOTAzIDM4Ljg0NDZWMzEuNjg3OEgxMDYuNjMxVjM4Ljc0MThDMTA2LjYzMSA0Mi41Mjk0IDEwNC41MjEgNDQuMzg2NCAxMDEuMTk2IDQ0LjM4NjRDOTcuODcwOCA0NC4zODY0IDk1LjgzMzUgNDIuNTE0NyA5NS44MzM1IDM4LjgyOTlaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik0xMDguNTE2IDMxLjY4NzhIMTExLjI0NFY0MS42ODUySDExNy40MjJWNDQuMTg4M0gxMDguNTE2VjMxLjY4NzhaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik0xMjIuODM1IDMxLjU5MjRIMTI1LjM1M0wxMzAuNjY0IDQ0LjE4MDlIMTI3LjgxMkwxMjYuNjc3IDQxLjM3NjlIMTIxLjQzOEwxMjAuMzAzIDQ0LjE4MDlIMTE3LjUyM0wxMjIuODM1IDMxLjU5MjRaTTEyNS43MDIgMzguOTQ3M0wxMjQuMDU3IDM0Ljg5NTVMMTIyLjQxMyAzOC45NDczSDEyNS43MDlIMTI1LjcwMloiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTEzMi45NzEgMzQuMjIwMkgxMjkuMjAyVjMxLjY4NzhIMTM5LjQ2OFYzNC4yMjAySDEzNS42OTlWNDQuMTg4M0gxMzIuOTcxVjM0LjIyMDJaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik0xMzkuNTg0IDM3Ljk3MTFWMzcuOTM0NEMxMzkuNTg0IDM0LjM4MTcgMTQyLjM2NCAzMS40Njc2IDE0Ni4xODQgMzEuNDY3NkMxNTAuMDA0IDMxLjQ2NzYgMTUyLjc1NCAzNC4zNDUgMTUyLjc1NCAzNy44OTc3VjM3LjkzNDRDMTUyLjc1NCA0MS40ODcgMTQ5Ljk3NSA0NC40MDExIDE0Ni4xNTUgNDQuNDAxMUMxNDIuMzM1IDQ0LjQwMTEgMTM5LjU4NCA0MS41MjM3IDEzOS41ODQgMzcuOTcxMVpNMTQ5LjkwOSAzNy45NzExVjM3LjkzNDRDMTQ5LjkwOSAzNS43OTEgMTQ4LjM1MiAzNC4wMDczIDE0Ni4xNTUgMzQuMDA3M0MxNDMuOTU3IDM0LjAwNzMgMTQyLjQzNyAzNS43NTQzIDE0Mi40MzcgMzcuODk3N1YzNy45MzQ0QzE0Mi40MzcgNDAuMDc3NyAxNDMuOTk0IDQxLjg2MTQgMTQ2LjE5MSA0MS44NjE0QzE0OC4zODkgNDEuODYxNCAxNDkuOTA5IDQwLjExNDQgMTQ5LjkwOSAzNy45NzExWiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNMTU0LjI5IDMxLjY4NzhIMTU5Ljk1OEMxNjEuNTI5IDMxLjY4NzggMTYyLjc1MiAzMi4xMzU2IDE2My41NjcgMzIuOTU3N0MxNjQuMjU4IDMzLjY1NSAxNjQuNjI5IDM0LjYzODYgMTY0LjYyOSAzNS44MTNWMzUuODQ5N0MxNjQuNjI5IDM3Ljg2ODMgMTYzLjU1MiAzOS4xMzgyIDE2MS45NzMgMzkuNzI1NEwxNjUgNDQuMTg4M0gxNjEuODEzTDE1OS4xNTcgNDAuMTg3OEgxNTcuMDE4VjQ0LjE4ODNIMTU0LjI5VjMxLjY4NzhaTTE1OS43NzYgMzcuNzU4MkMxNjEuMTA3IDM3Ljc1ODIgMTYxLjg2NCAzNy4wNDYyIDE2MS44NjQgMzUuOTg5MlYzNS45NTI1QzE2MS44NjQgMzQuNzcwNyAxNjEuMDQ5IDM0LjE2ODggMTU5LjcyNSAzNC4xNjg4SDE1Ny4wMThWMzcuNzU4MkgxNTkuNzc2WiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNNjAuMjk2OSA1LjQxNzFINjMuMjQzN0M2NC4wNzMyIDUuNDE3MSA2NC43MjggNS42NjY2NyA2NS4xNSA2LjA4NTA2QzY1LjQ5OTMgNi40NDQ3MyA2NS42OTU3IDYuOTQzODcgNjUuNjk1NyA3LjUyMzc1VjcuNTQ1NzdDNjUuNjk1NyA4LjYzOTQ2IDY1LjA0MDkgOS4zMDAwOSA2NC4xMzE0IDkuNTY0MzNMNjUuOTE0IDEyLjA4OTRINjQuNTM4OEw2Mi45MTYzIDkuNzYyNTJINjEuNDYxVjEyLjA4OTRINjAuMjk2OVY1LjQxNzFaTTYzLjE1NjQgOC43MjAyMUM2My45ODU5IDguNzIwMjEgNjQuNTE3IDguMjc5NzkgNjQuNTE3IDcuNjA0NDlWNy41ODI0N0M2NC41MTcgNi44NzA0NyA2NC4wMDc3IDYuNDgxNDMgNjMuMTQ5MSA2LjQ4MTQzSDYxLjQ2MVY4LjcyMDIxSDYzLjE2MzdINjMuMTU2NFoiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTY2LjYzNDMgOS41ODYzNlY5LjU2NDM0QzY2LjYzNDMgOC4xMjU2NSA2Ny42NDU3IDYuOTQzODcgNjkuMDcxOCA2Ljk0Mzg3QzcwLjY1OCA2Ljk0Mzg3IDcxLjQ1ODMgOC4xOTkwNSA3MS40NTgzIDkuNjUyNDJDNzEuNDU4MyA5Ljc1NTE4IDcxLjQ1MTEgOS44NjUyOSA3MS40MzY1IDkuOTc1MzlINjcuNzc2NkM2Ny45MDAzIDEwLjc5NzUgNjguNDc1MSAxMS4yNTI2IDY5LjIxIDExLjI1MjZDNjkuNzcwMyAxMS4yNTI2IDcwLjE2MzIgMTEuMDM5NyA3MC41NjM0IDEwLjY1MDdMNzEuMjMyOCAxMS4yNTI2QzcwLjc1OTggMTEuODI1MSA3MC4xMTIyIDEyLjE5MjEgNjkuMTk1NSAxMi4xOTIxQzY3Ljc0NzUgMTIuMTkyMSA2Ni42MzQzIDExLjEzNTEgNjYuNjM0MyA5LjU4NjM2Wk03MC4zMjMzIDkuMjEyMDFDNzAuMjUwNSA4LjQ3MDY0IDY5LjgxMzkgNy44OTA3NiA2OS4wNTcyIDcuODkwNzZDNjguMzU4NyA3Ljg5MDc2IDY3Ljg3MTIgOC40MzM5NCA2Ny43NjIxIDkuMjEyMDFINzAuMzIzM1oiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTcyLjU5MzYgMTMuMDIxNkw3My4wMTU2IDEyLjE1NTRDNzMuNTU0MSAxMi40ODU4IDc0LjExNDMgMTIuNjkxMyA3NC43OTEgMTIuNjkxM0M3NS44MDk3IDEyLjY5MTMgNzYuMzY5OSAxMi4xNTU0IDc2LjM2OTkgMTEuMTQ5OFYxMC43NjA4Qzc1Ljk1NTIgMTEuMzA0IDc1LjQzODYgMTEuNjg1NyA3NC42MTY0IDExLjY4NTdDNzMuNDQ0OSAxMS42ODU3IDcyLjM1MzUgMTAuODEyMiA3Mi4zNTM1IDkuMzI5NDVWOS4zMDc0M0M3Mi4zNTM1IDcuODI0NyA3My40NTk1IDYuOTQzODcgNzQuNjE2NCA2Ljk0Mzg3Qzc1LjQ1MzEgNi45NDM4NyA3NS45NzcgNy4zMzI5IDc2LjM2MjYgNy44MDI2OFY3LjA1Mzk3SDc3LjUwNVYxMS4wNTQ0Qzc3LjUwNSAxMS44OTg1IDc3LjI4NjcgMTIuNTI5OCA3Ni44NzIgMTIuOTQ4MkM3Ni40MjA5IDEzLjQwMzMgNzUuNzIyMyAxMy42MjM1IDc0LjgyMDEgMTMuNjIzNUM3My45OTc5IDEzLjYyMzUgNzMuMjMzOSAxMy40MTA2IDcyLjU5MzYgMTMuMDIxNlpNNzYuMzc3MiA5LjMxNDc3VjkuMjkyNzVDNzYuMzc3MiA4LjQ3MDY0IDc1LjcwNzggNy45MTI3OCA3NC45MTQ3IDcuOTEyNzhDNzQuMTIxNiA3LjkxMjc4IDczLjQ5NTkgOC40NjMzIDczLjQ5NTkgOS4yOTI3NVY5LjMxNDc3QzczLjQ5NTkgMTAuMTM2OSA3NC4xMjg5IDEwLjY5NDcgNzQuOTE0NyAxMC42OTQ3Qzc1LjcwMDUgMTAuNjk0NyA3Ni4zNzcyIDEwLjEyOTUgNzYuMzc3MiA5LjMxNDc3WiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNNzguOTM4IDUuMTg5NTVIODAuMTY3N1Y2LjI4MzI1SDc4LjkzOFY1LjE4OTU1Wk03OC45ODE2IDcuMDUzOTdIODAuMTI0VjEyLjA4Mkg3OC45ODE2VjcuMDUzOTdaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik04MS4xODY1IDExLjQyMTRMODEuNjk1OCAxMC42NDM0QzgyLjI0MTYgMTEuMDU0NCA4Mi44MTY0IDExLjI1OTkgODMuMzI1NyAxMS4yNTk5QzgzLjgzNSAxMS4yNTk5IDg0LjA5NyAxMS4wNDcxIDg0LjA5NyAxMC43MTY4VjEwLjY5NDdDODQuMDk3IDEwLjMwNTcgODMuNTY1OCAxMC4xNzM2IDgyLjk4MzcgOS45OTAwN0M4Mi4yNDg4IDkuNzc3MiA4MS40MjY2IDkuNDc2MjUgODEuNDI2NiA4LjUxNDY4VjguNDkyNjZDODEuNDI2NiA3LjUzODQzIDgyLjIxMjUgNi45NTg1NSA4My4yMDIgNi45NTg1NUM4My44Mjc3IDYuOTU4NTUgODQuNTA0NCA3LjE3ODc2IDg1LjAzNTYgNy41MzEwOUw4NC41ODQ1IDguMzUzMkM4NC4xMDQyIDguMDU5NTkgODMuNTk0OSA3Ljg3NjA4IDgzLjE4MDIgNy44NzYwOEM4Mi43MzYzIDcuODc2MDggODIuNDgxNyA4LjA4ODk1IDgyLjQ4MTcgOC4zNjc4OFY4LjM4OTlDODIuNDgxNyA4Ljc2NDI1IDgzLjAyMDEgOC45MTEwNSA4My42MDIyIDkuMTAxOUM4NC4zMjk4IDkuMzI5NDUgODUuMTUyIDkuNjY3MSA4NS4xNTIgMTAuNTY5OVYxMC41OTJDODUuMTUyIDExLjY0OSA4NC4zMzcxIDEyLjE3MDEgODMuMzAzOSAxMi4xNzAxQzgyLjU5ODEgMTIuMTcwMSA4MS44MTIzIDExLjkxMzIgODEuMTg2NSAxMS40MjE0WiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNODYuNTQ5IDEwLjY5NDdWOC4wNDQ5SDg1LjkxNlY3LjA1Mzk3SDg2LjU0OVY1LjY3NEg4Ny42OTE0VjcuMDUzOTdIODkuMDMwMlY4LjA0NDlIODcuNjkxNFYxMC41MTEyQzg3LjY5MTQgMTAuOTU5IDg3LjkxNjkgMTEuMTQyNSA4OC4zMDI2IDExLjE0MjVDODguNTU3MiAxMS4xNDI1IDg4Ljc4MjggMTEuMDgzOCA4OS4wMDg0IDEwLjk3MzdWMTEuOTEzMkM4OC43MjQ2IDEyLjA3NDcgODguNDExNyAxMi4xNzAxIDg3Ljk4MjQgMTIuMTcwMUM4Ny4xNDU3IDEyLjE3MDEgODYuNTQ5IDExLjc5NTggODYuNTQ5IDEwLjY5NDdaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik04OS43NzI1IDkuNTg2MzZWOS41NjQzNEM4OS43NzI1IDguMTI1NjUgOTAuNzgzOCA2Ljk0Mzg3IDkyLjIxIDYuOTQzODdDOTMuNzk2MSA2Ljk0Mzg3IDk0LjU5NjUgOC4xOTkwNSA5NC41OTY1IDkuNjUyNDJDOTQuNTk2NSA5Ljc1NTE4IDk0LjU4OTIgOS44NjUyOSA5NC41NzQ3IDkuOTc1MzlIOTAuOTE0OEM5MS4wMzg1IDEwLjc5NzUgOTEuNjEzMyAxMS4yNTI2IDkyLjM0ODIgMTEuMjUyNkM5Mi45MDg1IDExLjI1MjYgOTMuMzAxNCAxMS4wMzk3IDkzLjcwMTYgMTAuNjUwN0w5NC4zNzEgMTEuMjUyNkM5My44OTggMTEuODI1MSA5My4yNTA0IDEyLjE5MjEgOTIuMzMzNyAxMi4xOTIxQzkwLjg4NTcgMTIuMTkyMSA4OS43NzI1IDExLjEzNTEgODkuNzcyNSA5LjU4NjM2Wk05My40NjE0IDkuMjEyMDFDOTMuMzg4NyA4LjQ3MDY0IDkyLjk1MjEgNy44OTA3NiA5Mi4xOTU0IDcuODkwNzZDOTEuNDk2OSA3Ljg5MDc2IDkxLjAwOTQgOC40MzM5NCA5MC45MDAzIDkuMjEyMDFIOTMuNDYxNFoiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTk1Ljc2ODEgNy4wNTM5N0g5Ni45MTA0VjguMTg0MzdDOTcuMjIzMyA3LjQzNTY3IDk3Ljc5ODEgNi45MTQ1MSA5OC42Nzg1IDYuOTU4NTVWOC4xNzcwM0g5OC42MTNDOTcuNjA4OSA4LjE3NzAzIDk2LjkxMDQgOC44Mzc2NSA5Ni45MTA0IDEwLjE2NjJWMTIuMDgySDk1Ljc2ODFWNy4wNTM5N1oiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTk5LjMxODQgOS41ODYzNlY5LjU2NDM0Qzk5LjMxODQgOC4xMjU2NSAxMDAuMzMgNi45NDM4NyAxMDEuNzU2IDYuOTQzODdDMTAzLjM0MiA2Ljk0Mzg3IDEwNC4xNDIgOC4xOTkwNSAxMDQuMTQyIDkuNjUyNDJDMTA0LjE0MiA5Ljc1NTE4IDEwNC4xMzUgOS44NjUyOSAxMDQuMTIxIDkuOTc1MzlIMTAwLjQ2MUMxMDAuNTg0IDEwLjc5NzUgMTAxLjE1OSAxMS4yNTI2IDEwMS44OTQgMTEuMjUyNkMxMDIuNDU0IDExLjI1MjYgMTAyLjg0NyAxMS4wMzk3IDEwMy4yNDcgMTAuNjUwN0wxMDMuOTE3IDExLjI1MjZDMTAzLjQ0NCAxMS44MjUxIDEwMi43OTYgMTIuMTkyMSAxMDEuODggMTIuMTkyMUMxMDAuNDMyIDEyLjE5MjEgOTkuMzE4NCAxMS4xMzUxIDk5LjMxODQgOS41ODYzNlpNMTAzLjAwNyA5LjIxMjAxQzEwMi45MzUgOC40NzA2NCAxMDIuNDk4IDcuODkwNzYgMTAxLjc0MSA3Ljg5MDc2QzEwMS4wNDMgNy44OTA3NiAxMDAuNTU1IDguNDMzOTQgMTAwLjQ0NiA5LjIxMjAxSDEwMy4wMDdaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik0xMDUuMDQ1IDkuNTc5MDFWOS41NTY5OUMxMDUuMDQ1IDcuODkwNzYgMTA2LjE2NSA2Ljk1MTIxIDEwNy4zNTkgNi45NTEyMUMxMDguMTgxIDYuOTUxMjEgMTA4LjY5OCA3LjM2MjI2IDEwOS4wNjEgNy44MzkzOFY1LjEzODE3SDExMC4yMDRWMTIuMDg5NEgxMDkuMDYxVjExLjI1MjZDMTA4LjY5IDExLjc3MzcgMTA4LjE3NCAxMi4xOTIxIDEwNy4zNTkgMTIuMTkyMUMxMDYuMTg3IDEyLjE5MjEgMTA1LjA0NSAxMS4yNTI2IDEwNS4wNDUgOS41ODYzNlY5LjU3OTAxWk0xMDkuMDc2IDkuNTc5MDFWOS41NTY5OUMxMDkuMDc2IDguNTg4MDggMTA4LjM5OSA3Ljk0OTQ4IDEwNy42MjggNy45NDk0OEMxMDYuODU3IDcuOTQ5NDggMTA2LjE5NSA4LjU1ODcyIDEwNi4xOTUgOS41NTY5OVY5LjU3OTAxQzEwNi4xOTUgMTAuNTQ3OSAxMDYuODU3IDExLjE4NjUgMTA3LjYyOCAxMS4xODY1QzEwOC4zOTkgMTEuMTg2NSAxMDkuMDc2IDEwLjU0MDYgMTA5LjA3NiA5LjU3OTAxWiIgZmlsbD0iIzFEMUQxQiIvPgo8cGF0aCBkPSJNMTE0LjA1MyA3LjA1Mzk3SDExNS4yMjRMMTE2LjIyOCAxMC41MDM5TDExNy4zMzQgNy4wMzkyOUgxMTguMzE3TDExOS40MjIgMTAuNTAzOUwxMjAuNDQxIDcuMDUzOTdIMTIxLjU5MUwxMTkuOTYxIDEyLjExODdIMTE4LjkzNUwxMTcuODIyIDguNjgzNTFMMTE2LjcwMSAxMi4xMTg3SDExNS42ODNMMTE0LjA2NyA3LjA1Mzk3SDExNC4wNTNaIiBmaWxsPSIjMUQxRDFCIi8+CjxwYXRoIGQ9Ik0xMjIuNTIyIDUuMTg5NTVIMTIzLjc1MlY2LjI4MzI1SDEyMi41MjJWNS4xODk1NVpNMTIyLjU2NiA3LjA1Mzk3SDEyMy43MDhWMTIuMDgySDEyMi41NjZWNy4wNTM5N1oiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTEyNS4zODIgMTAuNjk0N1Y4LjA0NDlIMTI0Ljc0OVY3LjA1Mzk3SDEyNS4zODJWNS42NzRIMTI2LjUyNFY3LjA1Mzk3SDEyNy44NjNWOC4wNDQ5SDEyNi41MjRWMTAuNTExMkMxMjYuNTI0IDEwLjk1OSAxMjYuNzQ5IDExLjE0MjUgMTI3LjEzNSAxMS4xNDI1QzEyNy4zOSAxMS4xNDI1IDEyNy42MTUgMTEuMDgzOCAxMjcuODQxIDEwLjk3MzdWMTEuOTEzMkMxMjcuNTU3IDEyLjA3NDcgMTI3LjI0NCAxMi4xNzAxIDEyNi44MTUgMTIuMTcwMUMxMjUuOTc4IDEyLjE3MDEgMTI1LjM4MiAxMS43OTU4IDEyNS4zODIgMTAuNjk0N1oiIGZpbGw9IiMxRDFEMUIiLz4KPHBhdGggZD0iTTEyOS4wMiA1LjEzMDgzSDEzMC4xNjJWNy44MjQ3QzEzMC40ODIgNy4zNTQ5MiAxMzAuOTQ4IDYuOTUxMjEgMTMxLjcxOSA2Ljk1MTIxQzEzMi44MzkgNi45NTEyMSAxMzMuNDk0IDcuNzE0NTkgMTMzLjQ5NCA4Ljg3NDM1VjEyLjA4MkgxMzIuMzUyVjkuMjI2NjhDMTMyLjM1MiA4LjQ0ODYyIDEzMS45NjYgOC4wMDA4NiAxMzEuMjgyIDguMDAwODZDMTMwLjU5OCA4LjAwMDg2IDEzMC4xNjIgOC40NzA2NCAxMzAuMTYyIDkuMjQ4N1YxMi4wODk0SDEyOS4wMlY1LjEzMDgzWiIgZmlsbD0iIzFEMUQxQiIvPgo8L2c+CjxkZWZzPgo8Y2xpcFBhdGggaWQ9ImNsaXAwXzQ0NDZfNTI3MyI+CjxyZWN0IHdpZHRoPSIxNjUiIGhlaWdodD0iNTEiIGZpbGw9IndoaXRlIi8+CjwvY2xpcFBhdGg+CjwvZGVmcz4KPC9zdmc+Cg==" height="97px"/>
#> </a>
#> <address>Cancer Research UK is a registered charity in England and Wales (1089464), Scotland (SC041666), the Isle of Man (1103) and Jersey (247). A company limited by guarantee. Registered company in England and Wales (4325234) and the Isle of Man (5713F). Registered address: 2 Redman Place, London, E20 1JQ.</address>

if (interactive()) {
  crukFooter()
}